### 2023-06-06 #211 @whedon
[PRE REVIEW]: Check your outliers! An introduction to identifying statistical outliers in R with *easystats*
**Submitting author:** @rempsyc (<a href="http://orcid.org/0000-0003-4315-6788">Rémi Thériault</a>)
**Repository:** https://github.com/easystats/performance
**Branch with paper.md** (empty if default branch): JOSE_paper
**Version:**
