### 2023-06-06 #211 @rempsyc
Hi, as explained in the cover letter, the correct `.md` file is in the "JOSE_paper" branch (https://github.com/easystats/performance/tree/JOSE_paper). The submission form did not accept the branch URL and there was no box for specifying a branch. Is there a way to ask whedon to change the target repository? Thanks
