+++ papers/JOSE/mybibfile.bib
@@ -51,7 +51,7 @@
 
 @Article{ludecke2021performance,
     author = {Daniel Lüdecke and Mattan S. Ben-Shachar and Indrajeet Patil and Philip Waggoner and Dominique Makowski},
-    title = {{performance}: An R package for assessment, comparison and testing of statistical models},
+    title = {{performance}: An {R} package for assessment, comparison and testing of statistical models},
     volume = {6},
     number = {60},
     journal = {Journal of Open Source Software},
