### 2023-06-07 4459943
remove asterisk in affiliation
