### 2023-06-08 #211 @rempsyc
For potential reviewers, we would suggest authors of key cited papers in the current manuscript (including an expert statistician):
- Daniël Lakens (Lakens)
- Marie Delacre (mdelacre)
- Youri L. Mora (YouriLMora)

Else from the suggested list:
- mdogucu
- johnlray
- lebebr01
- nniiicc
- kbodwin
