### 2023-06-15 #211 @stats-tgeorge
Hello @rempsyc I am excited to start the review process of your submission! I looked over your materials and found a few items to get resolved before I invite reviews. See opened **issue 594** on your JOSE branch of the repository.
