### 2023-06-16 6e6a305
shorten paper [skip ci]
