+++ papers/JOSE/paper_longform.Rmd
@@ -47,6 +47,19 @@ simplesummary: |
   The *{performance}* package from the *easystats* ecosystem makes it easy to 
   diagnose outliers in R and according to current best practices thanks to the 
   `check_outiers()` function.
+abstract: |
+  Beyond the challenge of keeping up-to-date with current best practices 
+  regarding the diagnosis and treatment of outliers, an additional difficulty 
+  arises concerning the mathematical implementation of the recommended methods. 
+  In this paper, we provide an overview of current recommendations and best
+  practices and demonstrate how they can easily and conveniently be 
+  implemented in the R statistical computing software, using the *{performance}*
+  package of the *easystats* ecosystem. We cover univariate, multivariate, and 
+  model-based statistical outlier detection methods, their recommended 
+  threshold, standard output, and plotting methods. We conclude with 
+  recommendations on the handling of outliers: the different theoretical types 
+  of outliers, whether to exclude or winsorize them, and the importance of 
+    transparency.
 keywords: |
   univariate outliers; multivariate outliers; robust detection methods; R; easystats
 acknowledgement: |
@@ -103,10 +116,6 @@ library(see)
 library(datawizard)
 ```
 
-# Summary
-
-Beyond the challenge of keeping up-to-date with current best practices regarding the diagnosis and treatment of outliers, an additional difficulty arises concerning the mathematical implementation of the recommended methods. In this paper, we provide an overview of current recommendations and best practices and demonstrate how they can easily and conveniently be implemented in the R statistical computing software, using the *{performance}* package of the *easystats* ecosystem. We cover univariate, multivariate, and model-based statistical outlier detection methods, their recommended threshold, standard output, and plotting methods. We conclude with recommendations on the handling of outliers: the different theoretical types of outliers, whether to exclude or winsorize them, and the importance of transparency.
-
 # Statement of Need
 
 Real-life data often contain observations that can be considered *abnormal* when compared to the main population. The cause of it---be it because they belong to a different distribution (originating from a different generative process) or simply being extreme cases, statistically rare but not impossible---can be hard to assess, and the boundaries of "abnormal" difficult to define.
@@ -153,7 +162,22 @@ which(outliers)
 data_clean <- data[-which(outliers), ]
 ```
 
-All `check_outliers()` output objects possess a `plot()` method, meaning it is also possible to visualize the outliers using the generic `plot()` function on the resulting outlier object after loading the {see} package.
+All `check_outliers()` output objects possess a `plot()` method, meaning it is also possible to visualize the outliers:
+
+```{r univariate, eval=FALSE}
+library(see)
+
+plot(outliers)
+```
+
+```{r univariate_implicit, fig.cap = "Visual depiction of outliers using the robust z-score method. The distance represents an aggregate score for variables mpg, cyl, disp, and hp.", echo=FALSE}
+library(see)
+
+plot(outliers) + 
+  ggplot2::theme(axis.text.x = ggplot2::element_text(
+    angle = 45, size = 7
+))
+```
 
 Other univariate methods are available, such as using the interquartile range (IQR), or based on different intervals, such as the Highest Density Interval (HDI) or the Bias Corrected and Accelerated Interval (BCI). These methods are documented and described in the function's [help page](<https://easystats.github.io/performance/reference/check_outliers.html>).
 
@@ -163,13 +187,24 @@ Univariate outliers can be useful when the focus is on a particular variable, fo
 
 However, in many scenarios, variables of a data set are not independent, and an abnormal observation will impact multiple dimensions. For instance, a participant giving random answers to a questionnaire. In this case, computing the _z_ score for each of the questions might not lead to satisfactory results. Instead, one might want to look at these variables together.
 
[…]
