+++ papers/JOSE/paper.md
@@ -47,19 +47,6 @@ simplesummary: |
   The *{performance}* package from the *easystats* ecosystem makes it easy to 
   diagnose outliers in R and according to current best practices thanks to the 
   `check_outiers()` function.
-abstract: |
-  Beyond the challenge of keeping up-to-date with current best practices 
-  regarding the diagnosis and treatment of outliers, an additional difficulty 
-  arises concerning the mathematical implementation of the recommended methods. 
-  In this paper, we provide an overview of current recommendations and best
-  practices and demonstrate how they can easily and conveniently be 
-  implemented in the R statistical computing software, using the *{performance}*
-  package of the *easystats* ecosystem. We cover univariate, multivariate, and 
-  model-based statistical outlier detection methods, their recommended 
-  threshold, standard output, and plotting methods. We conclude with 
-  recommendations on the handling of outliers: the different theoretical types 
-  of outliers, whether to exclude or winsorize them, and the importance of 
-    transparency.
 keywords: |
   univariate outliers; multivariate outliers; robust detection methods; R; easystats
 acknowledgement: |
@@ -103,6 +90,10 @@ csl: apa.csl
 ---
 
 
+
+# Summary
+
+Beyond the challenge of keeping up-to-date with current best practices regarding the diagnosis and treatment of outliers, an additional difficulty arises concerning the mathematical implementation of the recommended methods. In this paper, we provide an overview of current recommendations and best practices and demonstrate how they can easily and conveniently be implemented in the R statistical computing software, using the *{performance}* package of the *easystats* ecosystem. We cover univariate, multivariate, and model-based statistical outlier detection methods, their recommended threshold, standard output, and plotting methods. We conclude with recommendations on the handling of outliers: the different theoretical types of outliers, whether to exclude or winsorize them, and the importance of transparency.
 
 # Statement of Need
 
@@ -186,18 +177,7 @@ which(outliers)
 data_clean <- data[-which(outliers), ]
 ```
 
-All `check_outliers()` output objects possess a `plot()` method, meaning it is also possible to visualize the outliers:
-
-
-```r
-library(see)
-
-plot(outliers)
-```
-
-\begin{figure}
-\includegraphics[width=1\linewidth]{paper_files/figure-latex/univariate_implicit-1} \caption{Visual depiction of outliers using the robust z-score method. The distance represents an aggregate score for variables mpg, cyl, disp, and hp.}\label{fig:univariate_implicit}
-\end{figure}
+All `check_outliers()` output objects possess a `plot()` method, meaning it is also possible to visualize the outliers using the generic `plot()` function on the resulting outlier object after loading the {see} package.
 
 Other univariate methods are available, such as using the interquartile range (IQR), or based on different intervals, such as the Highest Density Interval (HDI) or the Bias Corrected and Accelerated Interval (BCI). These methods are documented and described in the function's [help page](<https://easystats.github.io/performance/reference/check_outliers.html>).
 
@@ -207,7 +187,7 @@ Univariate outliers can be useful when the focus is on a particular variable, fo
 
 However, in many scenarios, variables of a data set are not independent, and an abnormal observation will impact multiple dimensions. For instance, a participant giving random answers to a questionnaire. In this case, computing the _z_ score for each of the questions might not lead to satisfactory results. Instead, one might want to look at these variables together.
 
[…]
