### 2023-06-22 d60f2eb
https://github.com/easystats/insight/issues/329
