+++ R/test_performance.R
@@ -32,11 +32,11 @@
 #' ## Nested vs. Non-nested Models
 #' Model's "nesting" is an important concept of models comparison. Indeed, many
 #' tests only make sense when the models are *"nested",* i.e., when their
-#' predictors are nested. This means that all the predictors of a model are
-#' contained within the predictors of a larger model (sometimes referred to as
-#' the encompassing model). For instance, `model1 (y ~ x1 + x2)` is
-#' "nested" within `model2 (y ~ x1 + x2 + x3)`. Usually, people have a list
-#' of nested models, for instance `m1 (y ~ 1)`, `m2 (y ~ x1)`,
+#' predictors are nested. This means that all the *fixed effects* predictors of
+#' a model are contained within the *fixed effects* predictors of a larger model
+#' (sometimes referred to as the encompassing model). For instance,
+#' `model1 (y ~ x1 + x2)` is "nested" within `model2 (y ~ x1 + x2 + x3)`. Usually,
+#' people have a list of nested models, for instance `m1 (y ~ 1)`, `m2 (y ~ x1)`,
 #' `m3 (y ~ x1 + x2)`, `m4 (y ~ x1 + x2 + x3)`, and it is conventional
 #' that they are "ordered" from the smallest to largest, but it is up to the
 #' user to reverse the order from largest to smallest. The test then shows
@@ -272,7 +272,7 @@ format.test_performance <- function(x, digits = 2, ...) {
 
   if (isTRUE(attributes(x)$is_nested)) {
     footer <- paste0(
-      "Models were detected as nested and are compared in sequential order.\n"
+      "Models were detected as nested (in terms of fixed parameters) and are compared in sequential order.\n"
     )
   } else {
     footer <- paste0(
