+++ pkgdown.yml
@@ -1,11 +1,11 @@
-pandoc: 3.1.2
+pandoc: 3.1.3
 pkgdown: 2.0.7.9000
 pkgdown_sha: 39083f38bbb2de2e1126529392bddf3e6e262ab4
 articles:
   check_model: check_model.html
   compare: compare.html
   r2: r2.html
-last_built: 2023-06-02T11:57Z
+last_built: 2023-06-22T06:51Z
 urls:
   reference: https://easystats.github.io/performance/reference
   article: https://easystats.github.io/performance/articles
