### 2023-07-14 #211 @stats-tgeorge
Hello @johnlray. I hope you are having a great summer! Would you be willing to review this JOSE submission?
