### 2023-07-23 5040b9b
Update check_model() defaults to align with https://github.com/easystats/see/pull/299
Also change default to `verbose = FALSE`
