### 2023-07-24 a0387ba
Always detrend
