+++ man/check_model.Rd
@@ -17,7 +17,7 @@ check_model(x, ...)
   dot_alpha = 0.8,
   colors = c("#3aaf85", "#1b6ca8", "#cd201f"),
   theme = "see::theme_lucid",
-  detrend = requireNamespace("qqplotr", quietly = TRUE),
+  detrend = TRUE,
   show_dots = NULL,
   bandwidth = "nrd",
   type = "density",
@@ -55,8 +55,8 @@ for dots, and third color for outliers or extreme values.}
 \item{theme}{String, indicating the name of the plot-theme. Must be in the
 format \code{"package::theme_name"} (e.g. \code{"ggplot2::theme_minimal"}).}
 
-\item{detrend}{Logical. Should Q-Q/P-P plots be de-trended?
-Defaults to \code{TRUE} if the \emph{qqplotr} package is installed and \code{FALSE} otherwise.}
+\item{detrend}{Logical. Should Q-Q/P-P plots be de-trended? Defaults to
+\code{TRUE}.}
 
 \item{show_dots}{Logical, if \code{TRUE}, will show data points in the plot. Set
 to \code{FALSE} for models with many observations, if generating the plot is too
