### 2023-07-31 80c04c5
Update check_model() defaults to align with see changes (#600)
Co-authored-by: Mattan S. Ben-Shachar <mattansb@msbstats.info>
