+++ man/check_normality.Rd
@@ -33,7 +33,7 @@ significant results for the distribution of residuals and visual inspection
 (e.g. Q-Q plots) are preferable. For generalized linear models, no formal
 statistical test is carried out. Rather, there's only a \code{plot()} method for
 GLMs. This plot shows a half-normal Q-Q plot of the absolute value of the
-standardized deviance residuals is shown (being in line with changes in
+standardized deviance residuals is shown (in line with changes in
 \code{plot.lm()} for R 4.3+).
 }
 \note{
