+++ DESCRIPTION
@@ -124,6 +124,7 @@ Suggests:
     patchwork,
     pscl,
     psych,
+    qqplotr (>= 0.0.6),
     randomForest,
     rmarkdown,
     rstanarm,
