### 2023-08-12 #221 @nniiicc
@stats-tgeorge - Not sure where to file this issue - but it seems that the `@editorialbot` is not responding in this repo ... See my command above, and also this example from another paper under review https://github.com/openjournals/jose-reviews/issues/209#issuecomment-1673165427
