### 2023-08-14 #221 @stats-tgeorge
I see you added a comment there. Sounds like a good plan. I thought it was worth 1 more try before adding to their workload. This repo situation is a little different since the review repo was created after editorialbot was implemented. Non-the-less is suspect the problems are related.
