+++ .github/SUPPORT.md
@@ -1,0 +1,29 @@
+# Getting help with `{performance}`
+
+Thanks for using `{performance}`. Before filing an issue, there are a few places
+to explore and pieces to put together to make the process as smooth as possible.
+
+Start by making a minimal **repr**oducible **ex**ample using the
+[reprex](http://reprex.tidyverse.org/) package. If you haven't heard of or used
+reprex before, you're in for a treat! Seriously, reprex will make all of your
+R-question-asking endeavors easier (which is a pretty insane ROI for the five to
+ten minutes it'll take you to learn what it's all about). For additional reprex
+pointers, check out the [Get help!](https://www.tidyverse.org/help/) resource
+used by the tidyverse team.
+
+Armed with your reprex, the next step is to figure out where to ask:
+
+  * If it's a question: start with StackOverflow. There are more people there to answer questions.
+  * If it's a bug: you're in the right place, file an issue.
+  * If you're not sure: let's [discuss](https://github.com/easystats/performance/discussions) it and try to figure it out! If your
+    problem _is_ a bug or a feature request, you can easily return here and
+    report it.
+
+Before opening a new issue, be sure to [search issues and pull requests](https://github.com/easystats/performance/issues) to make sure the
+bug hasn't been reported and/or already fixed in the development version. By
+default, the search will be pre-populated with `is:issue is:open`. You can
+[edit the qualifiers](https://help.github.com/articles/searching-issues-and-pull-requests/)
+(e.g. `is:pr`, `is:closed`) as needed. For example, you'd simply
+remove `is:open` to search _all_ issues in the repo, open or closed.
+
+Thanks for your help!
