### 2023-08-23 #221 @lebebr01
Hi @stats-tgeorge,

Yes I can, but I won't be able to get to it until early to mid September. Let me know if that timeline is too long.
