### 2023-08-25 #221 @stats-tgeorge
> Hi @stats-tgeorge,
>
> Yes I can, but I won't be able to get to it until early to mid September. Let me know if that timeline is too long.

@lebebr01 That works - we understand it is a busy time of year! Thank you!
