### 2023-08-25 15b609d
Bump `insight` version for `.get_dep_version` (#602)
