+++ DESCRIPTION
@@ -70,7 +70,7 @@ Depends:
     R (>= 3.6)
 Imports:
     bayestestR (>= 0.13.0),
-    insight (>= 0.19.1),
+    insight (>= 0.19.3.2),
     datawizard (>= 0.7.0),
     methods,
     stats,
@@ -148,3 +148,4 @@ Config/Needs/website:
     r-lib/pkgdown,
     easystats/easystatstemplate
 Config/rcmdcheck/ignore-inconsequential-notes: true
+Remotes: easystats/insight
