+++ pkgdown.yml
@@ -1,11 +1,11 @@
-pandoc: 3.1.6.1
+pandoc: 3.1.6.2
 pkgdown: 2.0.7.9000
 pkgdown_sha: c9206802f2888992de92aa41f517ba7812f05331
 articles:
   check_model: check_model.html
   compare: compare.html
   r2: r2.html
-last_built: 2023-08-21T07:00Z
+last_built: 2023-08-25T22:53Z
 urls:
   reference: https://easystats.github.io/performance/reference
   article: https://easystats.github.io/performance/articles
