+++ pkgdown.yml
@@ -5,7 +5,7 @@ articles:
   check_model: check_model.html
   compare: compare.html
   r2: r2.html
-last_built: 2023-08-25T22:53Z
+last_built: 2023-08-30T06:45Z
 urls:
   reference: https://easystats.github.io/performance/reference
   article: https://easystats.github.io/performance/articles
