### 2023-09-01 52a9b9e
minor printing fixes
