### 2023-09-04 db9190e
Add method for parsnip _glm model
