+++ NEWS.md
@@ -6,6 +6,9 @@
   to fixed effects parameters and currently ignores random effects when detecting
   nested models.
 
+* `check_outliers()` for `"ICS"` method is now more stable and less likely to
+  fail.
+
 # performance 0.10.4
 
 ## Changes to functions
