### 2023-09-10 342137c
test check_outliers ics which on R devel
