### 2023-09-10 8df9ad9
clean-up check_outliers tests
