### 2023-09-10 e5b2768
Collinearity not available for models of class 'hurdle' Fixes #606
