### 2023-09-10 eb8d4a1
Add check_convergence method for parsnip _glm model (#605)
Co-authored-by: Daniel <mail@danielluedecke.de>
