### 2023-09-10 4ebf4e4
Minor bug: Default method in item_intercor() is 'pearson', not 'spearman' Fixes #604
