+++ man/item_intercor.Rd
@@ -22,20 +22,19 @@ Compute various measures of internal consistencies
 for tests or item-scales of questionnaires.
 }
 \details{
-This function calculates a mean inter-item-correlation, i.e.
-a correlation matrix of \code{x} will be computed (unless
-\code{x} is already a matrix as returned by the \code{cor()}-function)
-and the mean of the sum of all item's correlation values is returned.
-Requires either a data frame or a computed \code{cor()}-object.
-\cr \cr
-\dQuote{Ideally, the average inter-item correlation for a set of
-items should be between .20 and .40, suggesting that while the
-items are reasonably homogeneous, they do contain sufficiently
-unique variance so as to not be isomorphic with each other.
-When values are lower than .20, then the items may not be
-representative of the same content domain. If values are higher than
-.40, the items may be only capturing a small bandwidth of the construct.}
-\cite{(Piedmont 2014)}
+This function calculates a mean inter-item-correlation, i.e. a
+correlation matrix of \code{x} will be computed (unless \code{x} is already a matrix
+as returned by the \code{cor()} function) and the mean of the sum of all items'
+correlation values is returned. Requires either a data frame or a computed
+\code{cor()} object.
+
+"Ideally, the average inter-item correlation for a set of items should be
+between 0.20 and 0.40, suggesting that while the items are reasonably
+homogeneous, they do contain sufficiently unique variance so as to not be
+isomorphic with each other. When values are lower than 0.20, then the items
+may not be representative of the same content domain. If values are higher
+than 0.40, the items may be only capturing a small bandwidth of the
+construct." \emph{(Piedmont 2014)}
 }
 \examples{
 data(mtcars)
