### 2023-09-10 e6e1721
Collinearity not available for models of class 'hurdle' (#609)
