+++ R/check_collinearity.R
@@ -440,6 +440,14 @@ check_collinearity.zerocount <- function(x,
 
   f <- insight::find_formula(x)
 
+  # hurdle or zeroinfl model can have no zero-inflation formula, in which case
+  # we have the same formula as for conditional formula part
+  if (inherits(x, c("hurdle", "zeroinfl", "zerocount")) &&
+      component == "zero_inflated" &&
+      is.null(f[["zero_inflated"]])) {
+    f$zero_inflated <- f$conditional
+  }
+
   if (inherits(x, "mixor")) {
     terms <- labels(x$terms)
   } else {
