+++ pkgdown.yml
@@ -5,7 +5,7 @@ articles:
   check_model: check_model.html
   compare: compare.html
   r2: r2.html
-last_built: 2023-09-10T20:39Z
+last_built: 2023-09-11T19:15Z
 urls:
   reference: https://easystats.github.io/performance/reference
   article: https://easystats.github.io/performance/articles
