### 2023-09-12 56e532a
bump to devel; fix styling and spelling workflows
