+++ NEWS.md
@@ -1,3 +1,5 @@
+# performance (development version)
+
 # performance 0.10.5
 
 ## Changes to functions
