### 2023-09-14 7ea53ed
avoid warning in example
