+++ R/binned_residuals.R
@@ -52,7 +52,7 @@
 #' \dontrun{
 #' # plot
 #' if (require("see")) {
-#'   plot(result)
+#'   plot(result, show_dots = TRUE)
 #' }
 #' }
 #'
