+++ man/check_convergence.Rd
@@ -63,31 +63,29 @@ or not.
 }
 
 \examples{
-if (require("lme4")) {
-  data(cbpp)
-  set.seed(1)
-  cbpp$x <- rnorm(nrow(cbpp))
-  cbpp$x2 <- runif(nrow(cbpp))
+\dontshow{if (require("lme4") && require("glmmTMB")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
+data(cbpp, package = "lme4")
+set.seed(1)
+cbpp$x <- rnorm(nrow(cbpp))
+cbpp$x2 <- runif(nrow(cbpp))
 
-  model <- glmer(
-    cbind(incidence, size - incidence) ~ period + x + x2 + (1 + x | herd),
-    data = cbpp,
-    family = binomial()
-  )
+model <- lm4::glmer(
+  cbind(incidence, size - incidence) ~ period + x + x2 + (1 + x | herd),
+  data = cbpp,
+  family = binomial()
+)
 
-  check_convergence(model)
-}
+check_convergence(model)
 
 \dontrun{
-if (require("glmmTMB")) {
-  model <- glmmTMB(
-    Sepal.Length ~ poly(Petal.Width, 4) * poly(Petal.Length, 4) +
-      (1 + poly(Petal.Width, 4) | Species),
-    data = iris
-  )
-  check_convergence(model)
+model <- suppressWarnings(glmmTMB::glmmTMB(
+  Sepal.Length ~ poly(Petal.Width, 4) * poly(Petal.Length, 4) +
+    (1 + poly(Petal.Width, 4) | Species),
+  data = iris
+))
+check_convergence(model)
 }
-}
+\dontshow{\}) # examplesIf}
 }
 \seealso{
 Other functions to check model assumptions and and assess model quality: 
