+++ R/icc.R
@@ -143,29 +143,25 @@
 #' very large, the variance ratio in the output makes no sense, e.g. because
 #' it is negative. In such cases, it might help to use `robust = TRUE`.
 #'
-#' @examples
-#' if (require("lme4")) {
-#'   model <- lmer(Sepal.Length ~ Petal.Length + (1 | Species), data = iris)
-#'   icc(model)
-#' }
+#' @examplesIf require("lme4")
+#' model <- lme4::lmer(Sepal.Length ~ Petal.Length + (1 | Species), data = iris)
+#' icc(model)
 #'
 #' # ICC for specific group-levels
-#' if (require("lme4")) {
-#'   data(sleepstudy)
-#'   set.seed(12345)
-#'   sleepstudy$grp <- sample(1:5, size = 180, replace = TRUE)
-#'   sleepstudy$subgrp <- NA
-#'   for (i in 1:5) {
-#'     filter_group <- sleepstudy$grp == i
-#'     sleepstudy$subgrp[filter_group] <-
-#'       sample(1:30, size = sum(filter_group), replace = TRUE)
-#'   }
-#'   model <- lmer(
-#'     Reaction ~ Days + (1 | grp / subgrp) + (1 | Subject),
-#'     data = sleepstudy
-#'   )
-#'   icc(model, by_group = TRUE)
+#' data(sleepstudy, package = "lme4")
+#' set.seed(12345)
+#' sleepstudy$grp <- sample(1:5, size = 180, replace = TRUE)
+#' sleepstudy$subgrp <- NA
+#' for (i in 1:5) {
+#'   filter_group <- sleepstudy$grp == i
+#'   sleepstudy$subgrp[filter_group] <-
+#'     sample(1:30, size = sum(filter_group), replace = TRUE)
 #' }
+#' model <- lme4::lmer(
+#'   Reaction ~ Days + (1 | grp / subgrp) + (1 | Subject),
+#'   data = sleepstudy
+#' )
+#' icc(model, by_group = TRUE)
 #' @export
 icc <- function(model,
                 by_group = FALSE,
