+++ R/check_multimodal.R
@@ -10,33 +10,29 @@
 #' @param x A numeric vector or a data frame.
 #' @param ... Arguments passed to or from other methods.
 #'
-#' @examples
+#' @examplesIf require("multimode") && require("mclust")
 #' \dontrun{
-#' if (require("multimode")) {
-#'   # Univariate
-#'   x <- rnorm(1000)
-#'   check_multimodal(x)
-#' }
+#' # Univariate
+#' x <- rnorm(1000)
+#' check_multimodal(x)
 #'
-#' if (require("multimode") && require("mclust")) {
-#'   x <- c(rnorm(1000), rnorm(1000, 2))
-#'   check_multimodal(x)
+#' x <- c(rnorm(1000), rnorm(1000, 2))
+#' check_multimodal(x)
 #'
-#'   # Multivariate
-#'   m <- data.frame(
-#'     x = rnorm(200),
-#'     y = rbeta(200, 2, 1)
-#'   )
-#'   plot(m$x, m$y)
-#'   check_multimodal(m)
+#' # Multivariate
+#' m <- data.frame(
+#'   x = rnorm(200),
+#'   y = rbeta(200, 2, 1)
+#' )
+#' plot(m$x, m$y)
+#' check_multimodal(m)
 #'
-#'   m <- data.frame(
-#'     x = c(rnorm(100), rnorm(100, 4)),
-#'     y = c(rbeta(100, 2, 1), rbeta(100, 1, 4))
-#'   )
-#'   plot(m$x, m$y)
-#'   check_multimodal(m)
-#' }
+#' m <- data.frame(
+#'   x = c(rnorm(100), rnorm(100, 4)),
+#'   y = c(rbeta(100, 2, 1), rbeta(100, 1, 4))
+#' )
+#' plot(m$x, m$y)
+#' check_multimodal(m)
 #' }
 #' @references
 #' - Ameijeiras-Alonso, J., Crujeiras, R. M., and Rodríguez-Casal, A. (2019).
