+++ man/check_distribution.Rd
@@ -45,12 +45,10 @@ implemented in the
 \href{https://easystats.github.io/see/}{\pkg{see}-package}.
 }
 \examples{
-if (require("lme4") && require("parameters") &&
-  require("see") && require("patchwork") && require("randomForest")) {
-  data(sleepstudy)
-
-  model <<- lmer(Reaction ~ Days + (Days | Subject), sleepstudy)
-  check_distribution(model)
-  plot(check_distribution(model))
-}
+\dontshow{if (require("lme4") && require("parameters") && require("see") && require("patchwork") && require("randomForest")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
+data(sleepstudy, package = "lme4")
+model <<- lme4::lmer(Reaction ~ Days + (Days | Subject), sleepstudy)
+check_distribution(model)
+plot(check_distribution(model))
+\dontshow{\}) # examplesIf}
 }
