+++ R/check_convergence.R
@@ -52,7 +52,7 @@
 #' cbpp$x <- rnorm(nrow(cbpp))
 #' cbpp$x2 <- runif(nrow(cbpp))
 #'
-#' model <- lm4::glmer(
+#' model <- lme4::glmer(
 #'   cbind(incidence, size - incidence) ~ period + x + x2 + (1 + x | herd),
 #'   data = cbpp,
 #'   family = binomial()
