### 2023-09-14 #221 @nniiicc
@stats-tgeorge - can you clarify that this is a learning module submission (as opposed to a software submission)?

I don't see any learning modules here to review. The repository connected to this submission https://github.com/easystats/performance is software. The only aspect of the repository that is related to learning or learning modules are contained in the JOSE paper ... Sorry if I am missing something obvious here.
