### 2023-09-14 #221 @stats-tgeorge
@rempsyc can you respond to this? @nniiicc would you consider it (2) open-source software, created as educational technology or infrastructure? We also accept that as part of JOSE. That was the type of submission I was considering it. I see the checklist is more specific to the learning modules though. I'm going to follow up to see if we have two versions. Thank you!
