### 2023-09-14 #221 @nniiicc
I think it could be either...

The paper is very strong - it gives a great overview of common dilemmas in identifying and overcoming issues related to outliers with the R package the authors have developed. The examples in the paper are clear, and have working code that I can run locally.

I guess stepping through the checklist I'm just confused what I should be reviewing... the paper, the software, or the learning modules (which are right now in the paper)...
