### 2023-09-14 #221 @rempsyc
We thought that open-source _software_ should be submitted to JOSS, whereas open-source _educational material_ should be submitted to JOSE.

This current submission is about educational material related to the proper use of the open-source software within the context of good statistical and research practice. What needs to be reviewed (the "module") is the content of the JOSE paper, not necessarily the software per se (the `performance` package has already been published in [JOSS](https://doi.org/10.21105/joss.03139) before). The current paper/module specifically focuses on the treatment of statistical outliers which has not been covered in the JOSS paper. Therefore, it seems that the reviewer checklist is not really adapted for aspects that target software components. Perhaps the software components can be safely ignored and the focus can remain on the JOSE paper.

Edit: so yes, in short, the module is contained within the paper itself
