+++ R/check_model.R
@@ -149,11 +149,6 @@
 #'   m <- lmer(Reaction ~ Days + (Days | Subject), sleepstudy)
 #'   check_model(m, panel = FALSE)
 #' }
-#'
-#' if (require("rstanarm")) {
-#'   m <- stan_glm(mpg ~ wt + gear, data = mtcars, chains = 2, iter = 200)
-#'   check_model(m)
-#' }
 #' }
 #' @export
 check_model <- function(x, ...) {
