+++ R/check_distribution.R
@@ -192,23 +192,23 @@ check_distribution.numeric <- function(model) {
   x <- x[!is.na(x)]
 
   data.frame(
-    "SD" = stats::sd(x),
-    "MAD" = stats::mad(x, constant = 1),
-    "Mean_Median_Distance" = mean(x) - stats::median(x),
-    "Mean_Mode_Distance" = mean(x) - as.numeric(bayestestR::map_estimate(x, bw = "nrd0")),
-    "SD_MAD_Distance" = stats::sd(x) - stats::mad(x, constant = 1),
-    "Var_Mean_Distance" = stats::var(x) - mean(x),
-    "Range_SD" = diff(range(x)) / stats::sd(x),
-    "Range" = diff(range(x)),
-    "IQR" = stats::IQR(x),
-    "Skewness" = .skewness(x),
-    "Kurtosis" = .kurtosis(x),
-    "Uniques" = length(unique(x)) / length(x),
-    "N_Uniques" = length(unique(x)),
-    "Min" = min(x),
-    "Max" = max(x),
-    "Proportion_Positive" = sum(x >= 0) / length(x),
-    "Integer" = all(.is_integer(x))
+    SD = stats::sd(x),
+    MAD = stats::mad(x, constant = 1),
+    Mean_Median_Distance = mean(x) - stats::median(x),
+    Mean_Mode_Distance = mean(x) - as.numeric(bayestestR::map_estimate(x, bw = "nrd0")),
+    SD_MAD_Distance = stats::sd(x) - stats::mad(x, constant = 1),
+    Var_Mean_Distance = stats::var(x) - mean(x),
+    Range_SD = diff(range(x)) / stats::sd(x),
+    Range = diff(range(x)),
+    IQR = stats::IQR(x),
+    Skewness = .skewness(x),
+    Kurtosis = .kurtosis(x),
+    Uniques = length(unique(x)) / length(x),
+    N_Uniques = length(unique(x)),
+    Min = min(x),
+    Max = max(x),
+    Proportion_Positive = sum(x >= 0) / length(x),
+    Integer = all(.is_integer(x))
   )
 }
 
