+++ R/r2_loo.R
@@ -20,9 +20,9 @@
 #'   leave-one-out-adjusted posterior distribution. This is conceptually similar
 #'   to an adjusted/unbiased R2 estimate in classical regression modeling. See
 #'   [r2_bayes()] for an "unadjusted" R2.
-#'   \cr \cr
+#'
 #'   Mixed models are not currently fully supported.
-#'   \cr \cr
+#'
 #'   `r2_loo_posterior()` is the actual workhorse for `r2_loo()` and
 #'   returns a posterior sample of LOO-adjusted Bayesian R2 values.
 #'
@@ -31,7 +31,9 @@
 #'
 #' @examples
 #' if (require("rstanarm")) {
-#'   model <- stan_glm(mpg ~ wt + cyl, data = mtcars, chains = 1, iter = 500, refresh = 0)
+#'   model <- suppressWarnings(
+#'     stan_glm(mpg ~ wt + cyl, data = mtcars, chains = 1, iter = 500, refresh = 0)
+#'   )
 #'   r2_loo(model)
 #' }
 #' @export
