+++ man/r2_bayes.Rd
@@ -64,7 +64,9 @@ returns a posterior sample of Bayesian R2 values.
 \examples{
 library(performance)
 if (require("rstanarm") && require("rstantools")) {
-  model <- stan_glm(mpg ~ wt + cyl, data = mtcars, chains = 1, iter = 500, refresh = 0)
+  model <- suppressWarnings(
+    stan_glm(mpg ~ wt + cyl, data = mtcars, chains = 1, iter = 500, refresh = 0)
+  )
   r2_bayes(model)
 
   model <- stan_lmer(
