+++ R/r2_bayes.R
@@ -33,9 +33,14 @@
 #' @examples
 #' library(performance)
 #' if (require("rstanarm") && require("rstantools")) {
-#'   model <- suppressWarnings(
-#'     stan_glm(mpg ~ wt + cyl, data = mtcars, chains = 1, iter = 500, refresh = 0)
-#'   )
+#'   model <- suppressWarnings(stan_glm(
+#'     mpg ~ wt + cyl,
+#'     data = mtcars,
+#'     chains = 1,
+#'     iter = 500,
+#'     refresh = 0,
+#'     show_messages = FALSE
+#'   ))
 #'   r2_bayes(model)
 #'
 #'   model <- stan_lmer(
