+++ man/r2_loo.Rd
@@ -53,9 +53,14 @@ returns a posterior sample of LOO-adjusted Bayesian R2 values.
 }
 \examples{
 if (require("rstanarm")) {
-  model <- suppressWarnings(
-    stan_glm(mpg ~ wt + cyl, data = mtcars, chains = 1, iter = 500, refresh = 0)
-  )
+  model <- suppressWarnings(stan_glm(
+    mpg ~ wt + cyl,
+    data = mtcars,
+    chains = 1,
+    iter = 500,
+    refresh = 0,
+    show_messages = FALSE
+  ))
   r2_loo(model)
 }
 }
