+++ man/check_model.Rd
@@ -206,15 +206,16 @@ skipped, which also increases performance.
 }
 
 \examples{
+\dontshow{if (require("lme4")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
 \dontrun{
 m <- lm(mpg ~ wt + cyl + gear + disp, data = mtcars)
 check_model(m)
 
-if (require("lme4")) {
-  m <- lmer(Reaction ~ Days + (Days | Subject), sleepstudy)
-  check_model(m, panel = FALSE)
+data(sleepstudy, package = "lme4")
+m <- lme4::lmer(Reaction ~ Days + (Days | Subject), sleepstudy)
+check_model(m, panel = FALSE)
 }
-}
+\dontshow{\}) # examplesIf}
 }
 \seealso{
 Other functions to check model assumptions and and assess model quality: 
