### 2023-09-14 3b390de
Use `examplesIf` in roxygen comments (#615)
