### 2023-09-15 6c95f67
try to reduce warnings in examples and use examplesIf
