### 2023-09-15 133cbe4
try to reduce warnings in examples and use examplesIf (#616)
