+++ R/looic.R
@@ -12,13 +12,13 @@
 #' @return A list with four elements, the ELPD, LOOIC and their standard errors.
 #'
 #' @examplesIf require("rstanarm")
-#' model <- rstanarm::stan_glm(
+#' model <- suppressWarnings(rstanarm::stan_glm(
 #'   mpg ~ wt + cyl,
 #'   data = mtcars,
 #'   chains = 1,
 #'   iter = 500,
 #'   refresh = 0
-#' )
+#' ))
 #' looic(model)
 #' @export
 looic <- function(model, verbose = TRUE) {
