+++ R/r2_bayes.R
@@ -43,13 +43,13 @@
 #'   ))
 #'   r2_bayes(model)
 #'
-#'   model <- stan_lmer(
+#'   model <- suppressWarnings(stan_lmer(
 #'     Petal.Length ~ Petal.Width + (1 | Species),
 #'     data = iris,
 #'     chains = 1,
 #'     iter = 500,
 #'     refresh = 0
-#'   )
+#'   ))
 #'   r2_bayes(model)
 #' }
 #'
@@ -75,10 +75,20 @@
 #'
 #' \dontrun{
 #' if (require("brms")) {
-#'   model <- brms::brm(mpg ~ wt + cyl, data = mtcars)
+#'   model <- suppressWarnings(brms::brm(
+#'     mpg ~ wt + cyl,
+#'     data = mtcars,
+#'     silent = 2,
+#'     refresh = 0
+#'   ))
 #'   r2_bayes(model)
 #'
-#'   model <- brms::brm(Petal.Length ~ Petal.Width + (1 | Species), data = iris)
+#'   model <- suppressWarnings(brms::brm(
+#'     Petal.Length ~ Petal.Width + (1 | Species),
+#'     data = iris,
+#'     silent = 2,
+#'     refresh = 0
+#'   ))
 #'   r2_bayes(model)
 #' }
 #' }
