+++ R/check_distribution.R
@@ -48,11 +48,14 @@ NULL
 #' There is a `plot()` method, which shows the probabilities of all predicted
 #' distributions, however, only if the probability is greater than zero.
 #'
-#' @examplesIf require("lme4") && require("parameters") && require("see") && require("patchwork") && require("randomForest")
+#' @examplesIf require("lme4") && require("parameters")
 #' data(sleepstudy, package = "lme4")
 #' model <<- lme4::lmer(Reaction ~ Days + (Days | Subject), sleepstudy)
 #' check_distribution(model)
+#'
+#' @examplesIf require("see") && require("patchwork") && require("randomForest")
 #' plot(check_distribution(model))
+#'
 #' @export
 check_distribution <- function(model) {
   UseMethod("check_distribution")
