+++ man/check_heteroscedasticity.Rd
@@ -28,7 +28,8 @@ This test of the hypothesis of (non-)constant error is also called
 \emph{Breusch-Pagan test} (\cite{1979}).
 }
 \note{
-There is also a \href{https://easystats.github.io/see/articles/performance.html}{\code{plot()}-method} implemented in the \href{https://easystats.github.io/see/}{\pkg{see}-package}.
+There is also a \href{https://easystats.github.io/see/articles/performance.html}{\code{plot()}-method}
+implemented in the \href{https://easystats.github.io/see/}{\pkg{see}-package}.
 }
 \examples{
 m <<- lm(mpg ~ wt + cyl + gear + disp, data = mtcars)
@@ -41,7 +42,8 @@ if (require("see")) {
 }
 }
 \references{
-Breusch, T. S., and Pagan, A. R. (1979) A simple test for heteroscedasticity and random coefficient variation. Econometrica 47, 1287-1294.
+Breusch, T. S., and Pagan, A. R. (1979) A simple test for heteroscedasticity
+and random coefficient variation. Econometrica 47, 1287-1294.
 }
 \seealso{
 Other functions to check model assumptions and and assess model quality: 
