### 2023-09-18 #221 @stats-tgeorge
@nniicc I verified with the EiC that we need the check all parts of the checklist since this was submitted (and appropriately so) as a learning module. Please continue as you suggested, reviewing the learning modules that are in the JOSE paper. It sounds like @rempsyc  cleared it up but I wanted to circle back. Let me know if there is more confusion and thank you!
