+++ R/performance_score.R
@@ -40,7 +40,7 @@
 #' model <- glm(counts ~ outcome + treatment, family = poisson())
 #'
 #' performance_score(model)
-#' \dontrun{
+#' \donttest{
 #' data(Salamanders, package = "glmmTMB")
 #' model <- glmmTMB::glmmTMB(
 #'   count ~ spp + mined + (1 | site),
