+++ R/check_normality.R
@@ -35,7 +35,7 @@
 #' x <- check_normality(m)
 #' plot(x)
 #'
-#' \dontrun{
+#' \donttest{
 #' # QQ-plot
 #' plot(check_normality(m), type = "qq")
 #'
