+++ R/r2_bayes.R
@@ -73,7 +73,7 @@
 #'   r2_bayes(model)
 #' }
 #'
-#' \dontrun{
+#' \donttest{
 #' if (require("brms")) {
 #'   model <- suppressWarnings(brms::brm(
 #'     mpg ~ wt + cyl,
