+++ R/r2_somers.R
@@ -8,7 +8,7 @@
 #' @return A named vector with the R2 value.
 #'
 #' @examples
-#' \dontrun{
+#' \donttest{
 #' if (require("correlation") && require("Hmisc")) {
 #'   model <- glm(vs ~ wt + mpg, data = mtcars, family = "binomial")
 #'   r2_somers(model)
