+++ R/check_outliers.R
@@ -299,7 +299,7 @@
 #' group_iris <- datawizard::data_group(iris, "Species")
 #' check_outliers(group_iris)
 #'
-#' \dontrun{
+#' \donttest{
 #' # You can also run all the methods
 #' check_outliers(data, method = "all")
 #'
