+++ R/check_convergence.R
@@ -60,7 +60,7 @@
 #'
 #' check_convergence(model)
 #'
-#' \dontrun{
+#' \donttest{
 #' model <- suppressWarnings(glmmTMB::glmmTMB(
 #'   Sepal.Length ~ poly(Petal.Width, 4) * poly(Petal.Length, 4) +
 #'     (1 + poly(Petal.Width, 4) | Species),
