+++ R/check_outliers.R
@@ -167,7 +167,9 @@
 #'  value for outliers classification. Refer to the help-file of
 #'  [ICSOutlier::ics.outlier()] to get more details about this procedure.
 #'  Note that `method = "ics"` requires both **ICS** and **ICSOutlier**
-#'  to be installed, and that it takes some time to compute the results.
+#'  to be installed, and that it takes some time to compute the results. You
+#'  can speed up computation time using parallel computing. Set the number of
+#'  cores to use with `options(mc.cores = 4)` (for example).
 #'
 #'  - **OPTICS**:
 #'  The Ordering Points To Identify the Clustering Structure (OPTICS) algorithm
@@ -1764,7 +1766,20 @@ check_outliers.metabin <- check_outliers.metagen
   n_cores <- if (!requireNamespace("parallel", quietly = TRUE)) {
     NULL
   } else {
-    max(1L, parallel::detectCores() - 2L, na.rm = TRUE)
+    getOption("mc.cores", 1L)
+  }
+
+  # tell user about n-cores option
+  if (is.null(n_cores)) {
+    insight::format_alert(
+      "Package `parallel` is not installed. `check_outliers()` will run on a single core.",
+      "Install package `parallel` and set, for example, `options(mc.cores = 4)` to run on multiple cores."
+    )
+  } else if (n_cores == 1) {
+    insight::format_alert(
+      "Package `parallel` is installed, but `check_outliers()` will run on a single core.",
+      "To use multiple cores, set `options(mc.cores = 4)` (for example)."
+    )
   }
 
   # Run algorithm
