### 2023-09-26 easystats/performance#619 @nniiicc
JOSE Review - 221: Statement of Need 
In the 'Statement of Need'
- Refer the reader to the github repository where they can find dependency information and installation instructions.
- Add one sentence about the intended audience for the instructional materials that follow. "Researchers" are clearly the broad audience, but at what level (undergraduate, graduate, professional, all of the above?)

https://github.com/openjournals/jose-reviews/issues/221
