### 2023-09-26 easystats/performance#620 @nniiicc
JOSE Review - 221: Community Contribution (to learning modules) 
The paper includes excellent instructions for how to identify and handle statistical outliers.

I can imagine someone working with researchers new to stats or easystats in R wanting to build off of your content for a classroom activity, or even an extended written tutorial.

To turn this content into something that could be reused I would suggest publishing the modules as a notebook, or a set of `.rmd` files. [Note: You may already have this content somewhere in the repo but I could not find it with a quick scan]

In the same directory where you place these files you can add a small note on how to contribute to this content.  This would allow me to check off two remaining boxes in my review:
- Usage: Does the documentation explain how someone would adopt the module, and include examples of how to use it?
- Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the module 2) Report issues or problems with the module 3) Seek support

https://github.com/openjournals/jose-reviews/issues/221
