### 2023-09-26 easystats/performance#621 @nniiicc
JOSE Review - 221: Previous Use / Instruction 
The final remaining box for me to tick is "Does it describe how it has been used in the classroom or other settings, and how someone might adopt it?" ... You can address this with a short line in the paper, or you can just tell me via the issue that this is not applicable to your submission.

https://github.com/openjournals/jose-reviews/issues/221
