### 2023-09-26 #221 @nniiicc
@rempsyc I have filed three small issues #619 #620 #621 in your repository. Overall, I really enjoyed the modules on outliers and look forward to getting this submission across the finish line soon.
