### 2023-09-27 easystats/performance#620 @bwiernik
@rempsyc
