### 2023-09-27 easystats/performance#619 @bwiernik
@rempsyc
