### 2023-09-27 easystats/performance#621 @bwiernik
@nniiicc if possible, could you please keep discussion of this paper on the issue in the JOSE repository? This repo is much broader and encompasses the full performance package and all of its features, not just outlier analysis
