### 2023-09-27 d459e1f
Recursion in `logLik.plm()` that slows down computations: Fixes #622
