### 2023-09-27 6325013
Update R/logLik.R
Co-authored-by: Indrajeet Patil <patilindrajeet.science@gmail.com>
