### 2023-09-27 fd61006
Update tests/testthat/test-logLik.R
Co-authored-by: Indrajeet Patil <patilindrajeet.science@gmail.com>
