### 2023-09-27 d15551b
Update tests/testthat/test-logLik.R
Co-authored-by: Indrajeet Patil <patilindrajeet.science@gmail.com>
