### 2023-09-27 201f3c3
Update tests/testthat/test-logLik.R
Co-authored-by: Indrajeet Patil <patilindrajeet.science@gmail.com>
