+++ .github/workflows/update-to-latest-easystats.yml
@@ -1,4 +1,4 @@
-name: Create PR for DESCRIPTION change
+name: Create PR for DESCRIPTION change to use latest easystats dependencies
 
 on:
   push:
@@ -7,8 +7,8 @@ on:
     branches: [main, master]
 
 # Run on a schedule (once a month)
-#schedule:
-#  - cron: "0 0 1 * *"
+# schedule:
+#   - cron: "0 0 1 * *"
 
 jobs:
   create_pr:
@@ -20,17 +20,11 @@ jobs:
       - name: Set up R
         uses: r-lib/actions/setup-r@v2
 
-      - name: Install dependencies
-        run: |
-          Rscript -e "install.packages('usethis')"
-        shell: bash
-        working-directory: ./ # Change this to the appropriate directory if necessary
-
       - name: Update DESCRIPTION file
         run: |
-          Rscript -e "usethis::use_latest_dependencies()"
-        shell: bash
-        working-directory: ./ # Change this to the appropriate directory if necessary
+          install.packages('usethis')
+          usethis::use_latest_dependencies()
+        shell: Rscript {0}
 
       - name: Check for changes in DESCRIPTION
         id: check_description
@@ -47,5 +41,5 @@ jobs:
           title: "Update DESCRIPTION to use latest 'easystats' dependencies"
           body: "Automatically updated the `DESCRIPTION` file."
           labels: "auto-update"
-          base: main # Change this to the target branch for your pull request
+          base: main
           branch-suffix: timestamp
