+++ .github/workflows/update-to-latest-easystats.yml
@@ -20,10 +20,14 @@ jobs:
       - name: Set up R
         uses: r-lib/actions/setup-r@v2
 
+      - uses: r-lib/actions/setup-r-dependencies@v2
+        with:
+          dependencies: "hard"
+          extra-packages: |
+            any::usethis
+
       - name: Update DESCRIPTION file
         run: |
-          install.packages('pak', dependencies = TRUE)
-          pak::pkg_install('usethis')
           usethis::use_latest_dependencies()
         shell: Rscript {0}
 
