+++ .github/workflows/update-to-latest-easystats.yml
@@ -39,11 +39,11 @@ jobs:
           git add --all
 
       - name: Create pull request
-        if: ${{ steps.update_description.outputs.changed }}
+        #if: ${{ steps.update_description.outputs.changed }}
         uses: peter-evans/create-pull-request@v5
         with:
           token: ${{ secrets.GITHUB_TOKEN }}
-          base: ${{ github.head_ref }}
+          base: main
           branch: desc-${{ github.ref_name }}-${{ github.job }}
           branch-suffix: timestamp
           delete-branch: true
