### 2023-09-27 4870b8c
Update `DESCRIPTION` to use latest 'easystats' dependencies (#625)
Co-authored-by: IndrajeetPatil <IndrajeetPatil@users.noreply.github.com>
