+++ DESCRIPTION
@@ -69,9 +69,9 @@ BugReports: https://github.com/easystats/performance/issues
 Depends:
     R (>= 3.6)
 Imports:
-    bayestestR (>= 0.13.0),
-    insight (>= 0.19.4),
-    datawizard (>= 0.7.0),
+    bayestestR (>= 0.13.1),
+    insight (>= 0.19.5),
+    datawizard (>= 0.9.0),
     methods,
     stats,
     utils
