### 2023-09-27 easystats/performance#621 @rempsyc
@nniiicc regarding this issue, I believe as you suggest that is not applicable to our submission, so closing this issue first. We have not had previous use of this module before this publication.
