### 2023-09-27 easystats/performance#619 @rempsyc
Fixed, thank you @nniiicc !
