### 2023-09-27 easystats/performance#620 @rempsyc
What I am thinking now is making a vignette out of it. We had already discussed this idea to make a more extended version. A vignette would also not be too much trouble and feat neatly on our website with other similar content.
