### 2023-09-27 eb41cd2
add check_outliers vignette
