### 2023-09-27 easystats/performance#620 @rempsyc
I have created a vignette with the extended paper and included a new section about reusing the educational module for teaching purposes at the beginning, which also includes a link to the community contribution guidelines.

It is available here: https://github.com/easystats/performance/blob/JOSE_paper/vignettes/check_outliers.Rmd

It is not up on the site yet as we still have to merge that branch on which we work on the JOSE paper. You can still knit the file locally to test it should you wish so. So closing this issue for now but of course let me know if there are any other issue.
