### 2023-09-29 5660f85
Error: `check_model()` not implemented for models of class `lm` yet Fixes #629
