+++ R/check_model_diagnostics.R
@@ -153,7 +153,7 @@
 # prepare data for normality of residuals plot ----------------------------------
 
 .diag_norm <- function(model, verbose = TRUE) {
-  r <- try(stats::residuals(model), silent = TRUE)
+  r <- try(as.numeric(stats::residuals(model)), silent = TRUE)
 
   if (inherits(r, "try-error")) {
     insight::format_alert(sprintf("Non-normality of residuals could not be computed. Cannot extract residuals from objects of class '%s'.", class(model)[1]))
@@ -212,8 +212,8 @@
   ncv <- tryCatch(
     {
       data.frame(
-        x = stats::fitted(model),
-        y = stats::residuals(model)
+        x = as.numeric(stats::fitted(model)),
+        y = as.numeric(stats::residuals(model))
       )
     },
     error = function(e) {
