### 2023-09-29 05336ec
Error: check_model() not implemented for models of class lm yet (#630)
