### 2023-10-02 #221 @stats-tgeorge
> In the eventuality that all reviewers are satisfied with the changes and that this paper gets accepted for publication before October 11, I will be able to include it in my postdoctoral fellowship application :) No worry if we cannot make this deadline.

Hello @rempsyc, I am working on this. If I cannot reconnect with the second reviewer soon I will seek another.
