### 2023-10-03 d1ede6a
Support for nestedLogit models
