+++ NAMESPACE
@@ -180,6 +180,7 @@ S3method(model_performance,model_fit)
 S3method(model_performance,multinom)
 S3method(model_performance,negbinirr)
 S3method(model_performance,negbinmfx)
+S3method(model_performance,nestedLogit)
 S3method(model_performance,plm)
 S3method(model_performance,poissonirr)
 S3method(model_performance,poissonmfx)
