+++ tests/testthat/_snaps/nestedLogit.md
@@ -1,0 +1,12 @@
+# model_performance
+
+    Code
+      model_performance(mnl)
+    Output
+      # Indices of model performance
+      
+      Response |     AIC |     BIC |  RMSE | Sigma
+      --------------------------------------------
+      work     | 325.733 | 336.449 | 0.456 | 1.000
+      full     | 110.495 | 118.541 | 0.398 | 1.000
+
