### 2023-10-03 easystats/performance#634 @lebebr01
JOSE Review - 221: Text Clarity
Overall, I enjoyed the JOSE paper. It was well-written and offered great insights.

For the review checklist element, "Instructional Design," a few small pieces are overstated in the JOSE paper. I would like to hear your thoughts on these.

1. pg 2, lines 43 - 46: There is a discussion of the mean and standard deviation not being robust, which is great. I was surprised to see the second part regarding these statistics assuming a Normal distribution. I agree and understand your point, but this is overstated for those learning statistics.
2. pg 7, lines 244 - 247: I liked this example and appreciate the idea of thinking about research context for extreme values/outliers. Would it be worth adding/framing this idea into statistical terms and being very explicit about what you mean by context?

https://github.com/openjournals/jose-reviews/issues/221
