### 2023-10-03 easystats/performance#635 @lebebr01
JOSE Review - 221: Contextualize Examples
One last element related to the "Instructional Design" and "Description" items for the JOSE review: I thought the JOSE paper would benefit from some contextualization of the analysis. For example, what are the goals of the analysis and context of the data? This can aid those trying to learn the material, helping them understand the context of the data and what the analysis is doing.

You could accomplish this in multiple ways. For example, one way could be a few sentences describing the example data at the beginning and some description of the results in the problem context.

https://github.com/openjournals/jose-reviews/issues/221
