### 2023-10-03 #221 @lebebr01
@rempsyc thank you for the module on outlier detection. I enjoyed reading it, and it offers a great example of outlier/extreme value detection. I filed two small issues in the main repository related to a few small points I thought were overstated in the JOSE paper and about adding context to the examples.
