### 2023-10-04 easystats/performance#635 @lebebr01
This is great added context. May seem superfluous given the usage of the data within R, but I really like the flow this adds to the paper.
