### 2023-10-04 easystats/performance#634 @lebebr01
Thanks! Looks great. The added piece to being explicit about the method's assumptions is helpful, I misread/misunderstood that.
