+++ NAMESPACE
@@ -372,6 +372,7 @@ S3method(r2,model_fit)
 S3method(r2,multinom)
 S3method(r2,negbinirr)
 S3method(r2,negbinmfx)
+S3method(r2,nestedLogit)
 S3method(r2,ols)
 S3method(r2,phylolm)
 S3method(r2,plm)
@@ -414,6 +415,7 @@ S3method(r2_coxsnell,mclogit)
 S3method(r2_coxsnell,multinom)
 S3method(r2_coxsnell,negbinirr)
 S3method(r2_coxsnell,negbinmfx)
+S3method(r2_coxsnell,nestedLogit)
 S3method(r2_coxsnell,poissonirr)
 S3method(r2_coxsnell,poissonmfx)
 S3method(r2_coxsnell,polr)
@@ -469,6 +471,7 @@ S3method(r2_nagelkerke,mclogit)
 S3method(r2_nagelkerke,multinom)
 S3method(r2_nagelkerke,negbinirr)
 S3method(r2_nagelkerke,negbinmfx)
+S3method(r2_nagelkerke,nestedLogit)
 S3method(r2_nagelkerke,poissonirr)
 S3method(r2_nagelkerke,poissonmfx)
 S3method(r2_nagelkerke,polr)
@@ -480,6 +483,8 @@ S3method(r2_posterior,BFBayesFactor)
 S3method(r2_posterior,brmsfit)
 S3method(r2_posterior,stanmvreg)
 S3method(r2_posterior,stanreg)
+S3method(r2_tjur,default)
+S3method(r2_tjur,nestedLogit)
 S3method(residuals,BFBayesFactor)
 S3method(residuals,check_normality_numeric)
 S3method(residuals,iv_robust)
