+++ tests/testthat/test-nestedLogit.R
@@ -1,4 +1,5 @@
 skip_on_os(c("mac", "linux"))
+skip_if(packageVersion("insight") <= "0.19.5.10")
 skip_if_not_installed("nestedLogit")
 skip_if_not_installed("carData")
 
