+++ tests/testthat/_snaps/check_distribution.md
@@ -1,0 +1,21 @@
+# check_distribution
+
+    Code
+      print(out)
+    Output
+      # Distribution of Model Family
+      
+      Predicted Distribution of Residuals
+      
+       Distribution Probability
+             cauchy         94%
+          lognormal          3%
+            weibull          3%
+      
+      Predicted Distribution of Response
+      
+        Distribution Probability
+           lognormal         47%
+               gamma         44%
+       beta-binomial          3%
+
