+++ tests/testthat/test-binned_residuals.R
@@ -30,7 +30,7 @@ test_that("binned_residuals, n_bins", {
   expect_equal(
     result$xbar,
     c(
-      0.02373, 0.06301, 0.08441, 0.17907, 0.29225, 0.44073, 0.54951, 
+      0.02373, 0.06301, 0.08441, 0.17907, 0.29225, 0.44073, 0.54951,
       0.69701, 0.9168, 0.99204
     ),
     tolerance = 1e-4
