+++ tests/testthat/test-check_autocorrelation.R
@@ -1,6 +1,7 @@
 test_that("check_autocorrelation", {
   data(mtcars)
   m <- lm(mpg ~ wt + cyl + gear + disp, data = mtcars)
+  set.seed(123)
   out <- check_autocorrelation(m)
-  expect_equal(out, 0.278, ignor_attr = TRUE, tolerance = 1e-3)
+  expect_equal(as.vector(out), 0.316, ignor_attr = TRUE, tolerance = 1e-2)
 })
