### 2023-10-04 32199f2
Support for nestedLogit models (#633)
