### 2023-10-04 easystats/performance#636 @stats-tgeorge
JOSE Review - 221
- [x] Line 38: change is to are
- [x] Lines 53-54: sentence could read better. Maybe something along the lines of: “identifying observations the model does not fit well”
- [x] It is not clear which regression models are supported regression models. If this is large list, is there a reference or a place the user could access the list? Can you add a sentence to make that clear?

Other comments (don't need resolved):
- Table 1 is very helpful; consider adding a column for the R function use. Right now it is very hard to read the thresholds column. Is the recommended thresholds the function defaults?
- Extending this to work well in the tidyverse or adding a section exhibiting that it already does might increase the adoption of your package. Consider compatibility with tidymodels specifically where the outlier detection can be part of the recipe for the data analysis.
- Great discussion on handling outliers (rather than just automatically removing them).

https://github.com/openjournals/jose-reviews/issues/221
