### 2023-10-04 44d11e1
addresses #636 major points
