### 2023-10-04 easystats/performance#636 @rempsyc
Thanks!

Point 1 has been corrected. Point 2 has been changed to:
> Identifying observations the regression model does not fit well can help find information relevant to our specific research context.

For point 3, I added,
> Currently, most lm models are supported (with the exception of `glmmTMB`, `lmrob`, and `glmrob` models), as long as they are supported by the underlying functions `stats::cooks.distance()` (or `loo::pareto_k_values()`) and `insight::get_data()` (for a full list of the 225 models currently supported by the `insight` package, see https://easystats.github.io/insight/#list-of-supported-models-by-class).

(addressing the other points in a second reply)
