### 2023-10-04 easystats/performance#636 @rempsyc
> Table 1 is very helpful; consider adding a column for the R function use. Right now it is very hard to read the thresholds column. Is the recommended thresholds the function defaults?

I clarified that the recommended thresholds are the default thresholds, but I find it difficult to format the table properly in rmarkdown PDF/JOSE paper. I tried using `"\n"` or adding empty rows to help air the content a bit but that does not seem to work.

Also, originally the table was larger but with four columns, the text starts overlapping, even when not using code formatting inside the table. I think it is because the function call (`check_outliers(model,`) is a single word with no spaces so it does not jump to the next line until the end of the string. I have for now removed the in-table code formatting in the hope of improving readability and added a column showing the function usage. See below for what I mean.

![table1](https://github.com/easystats/performance/assets/13123390/ed6aad0f-4460-4568-a483-f3702a217404)

I currently use `knitr::kable` but will try to see if other packages can yield a better outcome for the `.md` format.
