+++ papers/JOSE/paper.md
@@ -235,7 +235,78 @@ outliers
 
 Using the model-based outlier detection method, we identified a single outlier.
 
-All `check_outliers()` output objects possess a `plot()` method, meaning it is also possible to visualize the outliers using the generic `plot()` function on the resulting outlier object after loading the {see} package.
+Table 1 below summarizes which methods to use in which cases, and with what threshold. The recommended thresholds are the default thresholds.
+
+
+
+### Table 1
+
+_Summary of Statistical Outlier Detection Methods Recommendations_
+
+
+```{=latex}
+\global\setlength{\Oldarrayrulewidth}{\arrayrulewidth}
+
+\global\setlength{\Oldtabcolsep}{\tabcolsep}
+
+\setlength{\tabcolsep}{0pt}
+
+\renewcommand*{\arraystretch}{1.5}
+
+
+
+\providecommand{\ascline}[3]{\noalign{\global\arrayrulewidth #1}\arrayrulecolor[HTML]{#2}\cline{#3}}
+
+\begin{longtable}[c]{|p{1.25in}|p{1.25in}|p{1.25in}|p{1.25in}}
+
+
+
+\ascline{0.75pt}{000000}{1-4}
+
+\multicolumn{1}{>{\centering}m{\dimexpr 1.25in+0\tabcolsep}}{\textcolor[HTML]{000000}{\fontsize{10}{20}\selectfont{\global\setmainfont{Latin Modern Roman}{Statistical\ Test}}}} & \multicolumn{1}{>{\centering}m{\dimexpr 1.25in+0\tabcolsep}}{\textcolor[HTML]{000000}{\fontsize{10}{20}\selectfont{\global\setmainfont{Latin Modern Roman}{Diagnosis\ Method}}}} & \multicolumn{1}{>{\centering}m{\dimexpr 1.25in+0\tabcolsep}}{\textcolor[HTML]{000000}{\fontsize{10}{20}\selectfont{\global\setmainfont{Latin Modern Roman}{Recommended\ Threshold}}}} & \multicolumn{1}{>{\centering}m{\dimexpr 1.25in+0\tabcolsep}}{\textcolor[HTML]{000000}{\fontsize{10}{20}\selectfont{\global\setmainfont{Latin Modern Roman}{Function\ Usage}}}} \\
+
+\ascline{0.75pt}{000000}{1-4}\endfirsthead 
+
+\ascline{0.75pt}{000000}{1-4}
+
+\multicolumn{1}{>{\centering}m{\dimexpr 1.25in+0\tabcolsep}}{\textcolor[HTML]{000000}{\fontsize{10}{20}\selectfont{\global\setmainfont{Latin Modern Roman}{Statistical\ Test}}}} & \multicolumn{1}{>{\centering}m{\dimexpr 1.25in+0\tabcolsep}}{\textcolor[HTML]{000000}{\fontsize{10}{20}\selectfont{\global\setmainfont{Latin Modern Roman}{Diagnosis\ Method}}}} & \multicolumn{1}{>{\centering}m{\dimexpr 1.25in+0\tabcolsep}}{\textcolor[HTML]{000000}{\fontsize{10}{20}\selectfont{\global\setmainfont{Latin Modern Roman}{Recommended\ Threshold}}}} & \multicolumn{1}{>{\centering}m{\dimexpr 1.25in+0\tabcolsep}}{\textcolor[HTML]{000000}{\fontsize{10}{20}\selectfont{\global\setmainfont{Latin Modern Roman}{Function\ Usage}}}} \\
+
+\ascline{0.75pt}{000000}{1-4}\endhead
+
+
+
[…]
