### 2023-10-04 easystats/performance#636 @rempsyc
Ok, using the `flextable` package, I think we have a much better result now...

![new_table1](https://github.com/easystats/performance/assets/13123390/4e97f5b0-7a1c-4d2d-b1d0-737aaec8b3d1)
