+++ vignettes/check_outliers.Rmd
@@ -27,7 +27,8 @@ knitr::opts_chunk$set(
 )
 options(digits = 2)
 
-pkgs <- c("see", "performance", "datawizard", "rempsyc")
+pkgs <- c("see", "performance", "datawizard", "rempsyc", 
+          "ggplot2", "flextable", "ftExtra")
 successfully_loaded <- vapply(pkgs, requireNamespace, FUN.VALUE = logical(1L), quietly = TRUE)
 can_evaluate <- all(successfully_loaded)
 
@@ -154,7 +155,7 @@ Working with regression models creates the possibility of using model-based SOD 
 
 In {performance}, two such model-based SOD methods are currently available: Cook's distance, for regular regression models, and Pareto, for Bayesian models. As such, `check_outliers()` can be applied directly on regression model objects, by simply specifying `method = "cook"` (or `method = "pareto"` for Bayesian models).^[Our default threshold for the Cook method is defined by `stats::qf(0.5, ncol(x), nrow(x) - ncol(x))`, which again is an approximation of the critical value for _p_ < .001 consistent with the thresholds of our other methods.]
 
-Currently, most lm models are supported (with the exception of `glmmTMB`, `lmrob`, and `glmrob` models), as long as they are supported by the underlying functions `stats::cooks.distance()` (or `loo::pareto_k_values()`) and `insight::get_data()` (for a full list of the 225 models currently supported by the `insight` package, see https://easystats.github.io/insight/#list-of-supported-models-by-class). We show a demo below.
+Currently, most lm models are supported (with the exception of `glmmTMB`, `lmrob`, and `glmrob` models), as long as they are supported by the underlying functions `stats::cooks.distance()` (or `loo::pareto_k_values()`) and `insight::get_data()` (for a full list of the 225 models currently supported by the `insight` package, see https://easystats.github.io/insight/#list-of-supported-models-by-class). Also note that although `check_outliers()` supports the pipe operators (`|>` or `%>%`), it does not support `tidymodels` at this time. We show a demo below.
 
 ```{r model, fig.cap = "Visual depiction of outliers based on Cook's distance (leverage and standardized residuals), based on the fitted model."}
 model <- lm(disp ~ mpg * disp, data = data)
@@ -168,29 +169,39 @@ Using the model-based outlier detection method, we identified a single outlier.
 
 Table 1 below summarizes which methods to use in which cases, and with what threshold. The recommended thresholds are the default thresholds.
 
-```{r, echo=FALSE}
+```{r table1_prep, echo=FALSE}
 df <- data.frame(
   `Statistical Test` = c(
-    "Supported regression model",
-    "Structural Equation Modeling (or other unsupported model)",
-    "Simple test with few variables (*t* test, correlation, etc.)"
-  ),
+    "Supported regression model", 
+    "Structural Equation Modeling (or other unsupported model)", 
+    "Simple test with few variables (*t* test, correlation, etc.)"),
   `Diagnosis Method` = c(
-    "**Model-based**: Cook (or Pareto for Bayesian models)",
-    "**Multivariate**: Minimum Covariance Determinant (MCD)",
-    "**Univariate**: robust *z* scores (MAD)"
-  ),
+    "**Model-based**: Cook (or Pareto for Bayesian models)", 
+    "**Multivariate**: Minimum Covariance Determinant (MCD)", 
[…]
