+++ DESCRIPTION
@@ -93,7 +93,9 @@ Suggests:
     dbscan,
     estimatr,
     fixest,
+    flextable, 
     forecast,
+    ftExtra,
     gamm4,
     ggplot2,
     glmmTMB,
@@ -128,6 +130,7 @@ Suggests:
     psych,
     qqplotr (>= 0.0.6),
     randomForest,
+    rempsyc,
     rmarkdown,
     rstanarm,
     rstantools,
