### 2023-10-04 easystats/performance#636 @rempsyc
> Extending this to work well in the tidyverse or adding a section exhibiting that it already does might increase the adoption of your package. Consider compatibility with tidymodels specifically where the outlier detection can be part of the recipe for the data analysis.

I confirm that although `check_outliers()` works with the pipe and `insight::get_data()`, it does not currently work with tidymodels because of `stats::cooks.distance()`:

``` r
library(performance)

# Create some artificial outliers and an ID column
data <- rbind(mtcars[1:4], 42, 55)
data <- cbind(car = row.names(data), data)

lm(disp ~ mpg * hp, data = data) |>
  check_outliers() |>
  which()
#> [1] 31 34

suppressWarnings(library(tidymodels))

linear_reg() %>%
  fit(disp ~ mpg * hp, data = data) %>%
  check_outliers()
#> Error in UseMethod("cooks.distance"): no applicable method for 'cooks.distance' applied to an object of class "c('_lm', 'model_fit')"

# insight::get_data works
linear_reg() %>%
  fit(disp ~ mpg * hp, data = data) %>%
  insight::get_data()
#>                      disp  mpg  hp
#> Mazda RX4           160.0 21.0 110
#> Mazda RX4 Wag       160.0 21.0 110
#> Datsun 710          108.0 22.8  93
#> Hornet 4 Drive      258.0 21.4 110
#> Hornet Sportabout   360.0 18.7 175
#> Valiant             225.0 18.1 105

# stats::cooks.distance() doesn't
linear_reg() %>%
  fit(disp ~ mpg * hp, data = data) %>%
  stats::cooks.distance()
#> Error in UseMethod("cooks.distance"): no applicable method for 'cooks.distance' applied to an object of class "c('_lm', 'model_fit')"
```

I think it is worth thinking about adding support for those long-term, but for now, I have added `tidymodels` to the list of unsupported models within the paper, while mentioning that the pipe operator is supported.
> Also note that although `check_outliers()` supports the pipe operators (`|>` or `%>%`), it does not support `tidymodels` at this time.

With this I think I have addressed all the points so closing this issue for now but do not hesitate to comment again if more things come up. Thanks!
