### 2023-10-04 easystats/performance#636 @rempsyc
Even though
```r
output:
  rticles::joss_article:
    journal: "JOSE"
```
Renders the PDF correctly, it seems like the JOSE rendering process of the `.md` file does not like the LaTeX commands introduced by using `flextable` to produce the table... I suppose I'll just have to take a screenshot and leave it at that...
