+++ papers/JOSE/paper.Rmd
@@ -234,7 +234,7 @@ ftExtra::colformat_md(x)
 
 ```
 
-![]("paper_files/figure-latex/table1.jpg")
+![](table1.jpg)
 
 All `check_outliers()` output objects possess a `plot()` method, meaning it is also possible to visualize the outliers using the generic `plot()` function on the resulting outlier object after loading the {see} package (Figure 1).
 
