### 2023-10-05 #221 @rempsyc
Correct, I am waiting for permission to merge the JOSE branch in easystats/performance#586 and create a new release (since I am not the maintainer of the `performance` package) since this seems required by the instructions in order to put the release on Zenodo and get the doi. I'm going to wait a bit more and if we cannot merge the branch soon, I will simply build the package tar file and upload that on Zenodo. I was hesitant to do that because the `performance` package already [has a Zenodo](https://doi.org/10.5281/zenodo.8337481) that automatically updates with every release (which also changes the doi), so that will be a bit redundant. But creating a new release also means submitting the release to CRAN, which means other easystats dependencies have to be compatible, etc. Perhaps I can put the current branch on Zenodo as a separate file and get the doi and take care of the CRAN release later on? Perhaps @strengejacke can share his opinion on the best way to handle this as well.
