### 2023-10-05 #221 @rempsyc
@stats-tgeorge can you clarify whether the release and what should be on Zenodo should be of the `performance` package or of the educational module which is simply the paper??
