### 2023-10-05 187ab58
JOSE paper (#586)
Co-authored-by: Dominique Makowski <dom.mak19@gmail.com>
Co-authored-by: Daniel <mail@danielluedecke.de>
