+++ papers/JOSE/paper.bib
@@ -1,0 +1,161 @@
+@article{leys2019outliers,
+  title = {How to Classify, Detect, and Manage Univariate and Multivariate Outliers, With Emphasis on Pre-Registration},author = {Leys, Christophe and Delacre, Marie and Mora, Youri L. and Lakens, Daniël and Ley, Christophe},
+  journal = {International Review of Social Psychology},
+  year = {2019},
+  doi = {10.5334/irsp.289}
+}
+
+@article{leys2013outliers,
+  title = {Detecting outliers: Do not use standard deviation around the mean, use absolute deviation around the median},
+  author = {Christophe Leys and Christophe Ley and Olivier Klein and Philippe Bernard and Laurent Licata},
+  journal = {Journal of Experimental Social Psychology},
+  volume = {49},
+  number = {4},
+  pages = {764-766},
+  year = {2013},
+  doi = {10.1016/j.jesp.2013.03.013},
+  url = {https://doi.org/10.1016/j.jesp.2013.03.013}
+}
+
+@article{leys2018outliers,
+  title = {Detecting multivariate outliers: Use a robust variant of the Mahalanobis distance},
+  journal = {Journal of Experimental Social Psychology},
+  volume = {74},
+  pages = {150-156},
+  year = {2018},
+  issn = {0022-1031},
+  doi = {10.1016/j.jesp.2017.09.011},
+  url = {https://www.sciencedirect.com/science/article/pii/S0022103117302123},
+  author = {Christophe Leys and Olivier Klein and Yves Dominicy and Christophe Ley},
+}
+
+@article{simmons2011false,
+  author = {Joseph P. Simmons and Leif D. Nelson and Uri Simonsohn},
+  title ={False-Positive Psychology: Undisclosed Flexibility in Data Collection and Analysis Allows Presenting Anything as Significant},
+  journal = {Psychological Science},
+  volume = {22},
+  number = {11},
+  pages = {1359-1366},
+  year = {2011},
+  doi = {10.1177/0956797611417632},
+  URL = {https://doi.org/10.1177/0956797611417632},
+}
+
+@software{easystatspackage,
+  title = {{easystats}: Streamline Model Interpretation, Visualization, and Reporting},
+  author = {Daniel Lüdecke and Dominique Makowski and Mattan S. Ben-Shachar and Indrajeet Patil and Brenton M. Wiernik and Etienne Bacher and Rémi Thériault},
+  date = {2023-02-04T22:06:06Z},
+  origdate = {2019-01-28T10:39:29Z},
+  url = {https://easystats.github.io/easystats/}
+}
+
+@Article{ludecke2021performance,
+    author = {Daniel Lüdecke and Mattan S. Ben-Shachar and Indrajeet Patil and Philip Waggoner and Dominique Makowski},
+    title = {{performance}: An {R} package for assessment, comparison and testing of statistical models},
+    volume = {6},
+    number = {60},
+    journal = {Journal of Open Source Software},
+    year = {2021},
+    pages = {3139},
+    doi = {10.21105/joss.03139},
+    url = {https://doi.org/10.21105/joss.03139}
+  }
+
+@Article{patil2022datawizard,
+    title = {{datawizard}: An {R} package for easy data preparation and statistical transformations},
+    author = {Indrajeet Patil and Dominique Makowski and Mattan S. Ben-Shachar and Brenton M. Wiernik and Etienne Bacher and Daniel Lüdecke},
+    journal = {Journal of Open Source Software},
+    year = {2022},
+    volume = {7},
+    number = {78},
+    pages = {4684},
+    doi = {10.21105/joss.04684},
+  }
+
+@article{cook1977detection,
[…]
