+++ vignettes/paper.bib
@@ -41,18 +41,17 @@
   URL = {https://doi.org/10.1177/0956797611417632},
 }
 
-@Article{easystatspackage,
-    title = {easystats: Framework for Easy Statistical Modeling, Visualization, and Reporting},
-    author = {Daniel Lüdecke and Mattan S. Ben-Shachar and Indrajeet Patil and Brenton M. Wiernik and Etienne Bacher and Rémi Thériault and Dominique Makowski},
-    journal = {CRAN},
-    year = {2022},
-    note = {R package},
-    url = {https://easystats.github.io/easystats/},
-  }
+@software{easystatspackage,
+  title = {{easystats}: Streamline Model Interpretation, Visualization, and Reporting},
+  author = {Daniel Lüdecke and Dominique Makowski and Mattan S. Ben-Shachar and Indrajeet Patil and Brenton M. Wiernik and Etienne Bacher and Rémi Thériault},
+  date = {2023-02-04T22:06:06Z},
+  origdate = {2019-01-28T10:39:29Z},
+  url = {https://easystats.github.io/easystats/}
+}
 
 @Article{ludecke2021performance,
     author = {Daniel Lüdecke and Mattan S. Ben-Shachar and Indrajeet Patil and Philip Waggoner and Dominique Makowski},
-    title = {{performance}: An R package for assessment, comparison and testing of statistical models},
+    title = {{performance}: An {R} package for assessment, comparison and testing of statistical models},
     volume = {6},
     number = {60},
     journal = {Journal of Open Source Software},
@@ -146,7 +145,7 @@
 }
 
 @book{mcelreath2020statistical,
-  title={Statistical rethinking: A Bayesian course with examples in R and Stan},
+  title={Statistical rethinking: A Bayesian course with examples in {R} and Stan},
   author={McElreath, Richard},
   year={2020},
   publisher={CRC press}
