### 2023-10-05 #221 @rempsyc
Thanks @stats-tgeorge

- The release for the latest version, v0.10.6 (which contains the package, the paper, and the new vignette) is available here: https://github.com/easystats/performance/releases/tag/v0.10.6
- The release has been archived on Zenodo: https://doi.org/10.5281/zenodo.8411009

I think I've double-checked everything and checked all the boxes on my end. However, I am not able to check the boxes on your post myself, as the post is locked for me.
