### 2023-10-05 #221 @stats-tgeorge
@rempsyc

- [ ] The archive title has to match the paper title.
- [ ] Author list needs to match paper author list
- [ ] License in archive just links back to archive
