### 2023-10-05 #221 @stats-tgeorge
@rempsyc  two reference in your paper are missing DOIs. Can you find those?
