### 2023-10-05 1c68048
add doi for Gnanadesikan 1972 and render paper
