+++ papers/JOSE/paper.bib
@@ -94,6 +94,7 @@
 @article{gnanadesikan1972robust,
   title = {Robust estimates, residuals, and outlier detection with multiresponse data},
   author = {Gnanadesikan, R. and Kettenring, J. R},
+  doi = {10.2307/2528963},
   journal = {Biometrics},
   pages = {81-124},
   year = {1972}
