### 2023-10-05 #221 @rempsyc
> two reference in your paper are missing DOIs. Can you find those?

- [x] Gnanadesikan 1972: added
- [ ] Tukey 1963: [no doi available](https://www.jstor.org/stable/25049278)
