### 2023-10-05 #221 @rempsyc
Thanks for the clarification!

> - [x] The archive title has to match the paper title.
> - [x]  Author list needs to match paper author list
> - [x]  License in archive just links back to archive

All fixed now. New doi: [10.5281/zenodo.8411224](https://doi.org/10.5281/zenodo.8411224)
