### 2023-10-05 8ef0626
avoid loaded term
