+++ R/test_wald.R
@@ -10,7 +10,7 @@ test_wald.default <- function(..., verbose = TRUE) {
   # Attribute class to list and get names from the global environment
   objects <- insight::ellipsis_info(..., only_models = TRUE)
 
-  # Sanity checks (will throw error if non-valid objects)
+  # validation checks (will throw error if non-valid objects)
   objects <- .test_performance_checks(objects, verbose = verbose)
 
   # ensure proper object names
