+++ pkgdown.yml
@@ -6,7 +6,7 @@ articles:
   check_outliers: check_outliers.html
   compare: compare.html
   r2: r2.html
-last_built: 2023-10-05T16:18Z
+last_built: 2023-10-05T17:21Z
 urls:
   reference: https://easystats.github.io/performance/reference
   article: https://easystats.github.io/performance/articles
