### 2023-10-05 #221 @editorialbot
:wave: @openjournals/jose-eics, this paper is ready to be accepted and published.
