### 2023-10-06 easystats/performance#620 @DominiqueMakowski
nice one congrats!
