+++ R/model_performance.rma.R
@@ -18,7 +18,7 @@
 #'
 #'   - **AIC** Akaike's Information Criterion, see `?stats::AIC`
 #'
-#'   - **BIC** {Bayesian Information Criterion, see `?stats::BIC`}
+#'   - **BIC** Bayesian Information Criterion, see `?stats::BIC`
 #'
 #'    - **I2**: For a random effects model, `I2` estimates (in
 #'    percent) how much of the total variability in the effect size estimates
