+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Type: Package
 Package: performance
 Title: Assessment of Regression Models Performance
-Version: 0.10.6
+Version: 0.10.6.1
 Authors@R: 
     c(person(given = "Daniel",
              family = "Lüdecke",
@@ -153,4 +153,3 @@ Config/Needs/website:
     r-lib/pkgdown,
     easystats/easystatstemplate
 Config/rcmdcheck/ignore-inconsequential-notes: true
-Remotes: easystats/see, easystats/parameters, easystats/insight
