+++ tests/testthat/test-check_autocorrelation.R
@@ -3,5 +3,5 @@ test_that("check_autocorrelation", {
   m <- lm(mpg ~ wt + cyl + gear + disp, data = mtcars)
   set.seed(123)
   out <- check_autocorrelation(m)
-  expect_equal(as.vector(out), 0.316, ignor_attr = TRUE, tolerance = 1e-2)
+  expect_equal(as.vector(out), 0.316, ignore_attr = TRUE, tolerance = 1e-2)
 })
