### 2023-10-17 #221 @labarba
On a first browse of this submission, I am confused. The repository link points to the package `performance`, which has already been published in JOSS. So where is the educational material we are publishing in JOSE?
