### 2023-10-17 #221 @stats-tgeorge
> I am looking through the history of this submission, and found the first [review checklist](https://github.com/openjournals/jose-reviews/issues/221#issuecomment-1682722467), showing that this was submitted as a "learning module." (JOSE accepts two kinds of papers: those reporting on learning modules, and those reporting on educational software.)
>
> However, the associated repository that has been linked to this submission points to the _software package_ `performance`—a piece of software already reviewed and published in JOSS.
>
> Do I take it that the authors have written their "learning module" **in the paper itself?** And there is no associated "learning module" to go along with this submission?

That is correct; the learning module is within the paper. The paper was in the JOSE branch via the repo link at the top but in order for the authors to create a zenodo archive they had to reorganize. The repo for performance still has a folder with the JOSE paper in it.
