### 2023-10-17 #221 @labarba
Yikes. I'm afraid this breaks the JOSE model completely.

We can't link the paper to the repository of the software—it is a scholarly output that was already published in JOSS.
I'm not really sure what to do here, given the effort everyone has already invested.

I apologize to the authors and reviewers that this was not flagged earlier in the process.

The essential philosophy of JOSE is one that matches JOSS, where papers report on separate scholarly artifacts that are traditionally not rewarded with publication in the prevailing system. But in JOSE, the papers are _about_ open source teaching materials _that are intended to be reused or modified_ in the open source model, or about educational software.

We do not publish papers that are themselves, in their narrative, intended to be the lesson material. This should be clear in the [documentation](https://openjournals.readthedocs.io/en/jose/), but if it is not, please help us see where the confusion arises.
