### 2023-10-17 #221 @rempsyc
We have the full educational module (i.e., an extended version of the paper) as a vignette: https://easystats.github.io/performance/articles/check_outliers.html

It includes a statement about reuse of the material and instructions on how to contribute to the educational module. The repo of the vignette is still in the `performance` package repository though. Would you like us to create a new repo just for the module? In that case the issue is that the vignette would not be updated automatically with our website. Perhaps we can simply link to the vignette or the [vignette source file](https://github.com/easystats/performance/blob/HEAD/vignettes/check_outliers.Rmd) somehow?
