### 2023-10-17 #221 @bwiernik
I'm not really following how the vignette for this paper living within a much larger software repo contradicts the mission of the JOSE to support open source educational materials.

Nothing about the current outliers paper was previously published or recognized by JOSS.

If it would be preferred, we can move the outliers vignette to its own repository.
