### 2023-10-18 #221 @nniiicc
@labarba I think there is some confusion about the review process we've already been through and the vingettes that @rempsyc points to above... TLDR - please give advice on Issues below

> We can't link the paper to the repository of the software—it is a scholarly output that was already published in JOSS.
> I'm not really sure what to do here, given the effort everyone has already invested.
> I apologize to the authors and reviewers that this was not flagged earlier in the process.

See https://github.com/openjournals/jose-reviews/issues/221#issuecomment-1719813372 - this issue was raised, and @stats-tgeorge consulted an EiC

There seems to be three issues:

1. Repo linked to JOSS paper
2. 1:1 mapping between *unique repo with only OER material* and the paper

@bwiernik suggestion to move the vignettes to a new repo should solve both issues 1 and 2

3. Vignettes as OER: If the issue is that the vignettes are not a *standalone* contributions recognized by JOSE as an OER then I think both the authors and reviewers were given poor directions about what constitutes an OER module submission.

If 3 is really an issue for JOSE editors - **I recommend for the sake of good will and for seeing @rempsyc's hard work rewarded that we give some very clear directions on how to extend the vignettes into an appropriate OER submission, and how he should revise the actual JOSE paper accordingly.**

*I realize that giving this kind of feedback (in Issue 3) might not be what an EiC commonly does w/r/t a JOSE submission - but I want to advocate strongly on the lead authors behalf here -  he wrote out these modules seeking credit for educational work he is doing with a statistical package he helped develop for a postdoc application he is applying to....The form or structure of the submission aside, this scenario is exactly what I *believe* JOSE was designed to address. If we can't offer an early-career computational scientist with a venue to document their open educational statistical work - that IMHO breaks the JOSE model completely
