### 2023-10-18 #221 @labarba
Again, I apologize to the authors and reviewers for the confusion. I take full responsibility for this as EiC. As a way of explanation, when a new submission comes in, I inspect it with the goal of finding the best fit of handling editor. This is often complicated by having to balance the load across all editors. Given my own workload, this inspection cannot be deep, so I missed that: (1) the linked repo was to an already published software artifact; (2) the learning material was contained in the paper itself.

The issues above should have been raised in the review process, however. It looks like they were not flagged then, either. I acknowledge that the handling editor (@stats-tgeorge) did message me in Slack about this submission, and I now have found that he wrote "Their learning content is within the JOSE paper" within a longer message—but again _I missed this_.

As a side note: you will notice that the editorial bot crawls the repository, and reports statistics. We use these to inform an assessment of the scale of scholarly work. In this case, they were reporting on the full software repo, and thus gave an incorrect signal of the work being submitted for review here.

From the [comment above](https://github.com/openjournals/jose-reviews/issues/221#issuecomment-1767320684), I see that you have crafted a "vignette," which is written in R markdown, hosted in your software repository, and can be [read online](https://easystats.github.io/performance/articles/check_outliers.html). THAT should have been the material reviewed, and about which the paper (briefly) reports. This seems to be another boo-boo in the review process here, namely, the paper is too long ([editorialbot reports](https://github.com/openjournals/jose-reviews/issues/221#issuecomment-1677559810) 3428 words). JOSE papers should be short and contain specific sections. From the [documentation](https://openjournals.readthedocs.io/en/jose/submitting.html#what-should-my-paper-contain):

> JOSE papers contain a limited set of metadata, plus Summary & Reference sections. We explicitly do not publish long-form articles, because the scholarship represented by a JOSE article is contained in the software or learning modules themselves. Expected length is around 1000 words max.

In sum, the educational content the paper is about should exist in an open repository, while [the paper contains](https://openjournals.readthedocs.io/en/jose/submitting.html#what-should-my-paper-contain) only specific items of description.

When assessing the scholarly contribution of the learning content, one criterion is the length, as measured by the editorialbot stats on the repo. Therefore, it is helpful if the repo is not bloated by extraneous materials. One solution can be to make a branch where you remove all other material and leave the vignette only, and we use that branch of the repo to link to the JOSE submission.

The Zenodo archive should also be of the learning content as hosted in its source repository. Note that the JOSE publishing process makes a Crossref deposit where the DOI of the paper includes the DOI of the archive as a linked artifact via the metadata.

The paper itself will have to be cut down to just the requested sections, with the learning content itself residing solely in the vignette. I do have additional concerns about the length of the vignette, and point you to the [documentation](https://openjournals.readthedocs.io/en/jose/review_criteria.html#specific-requirements-for-learning-modules), where we state:

> A learning module should cover a substantial portion of material to achieve plainly clear learning objectives. The ideal module consists of a few lessons,building up a well-rounded topic, as a full tutorial or part of a term or semester course.

JOSE _has_ published papers on shorter learning modules, and the editors have discussed several times our varying opinions on what constitutes substantial scholarly contribution. We don't all agree, and we accept that length by itself is an imperfect criterion. Our determination was that shorter lessons can merit publication when they are supported by other aspects of scholarly contribution, e.g., if the lesson has been taught and has been assessed in the classroom.

I'll stay tuned here for your reactions or comments and commit to a constructive approach to resolving the muddle.
