+++ papers/preprint/paper.Rmd
@@ -9,22 +9,22 @@ tags:
 authors:
   - name: Rémi Thériault
     orcid: 0000-0003-4315-6788
-    affiliation: 1
+    affiliation: Department of Psychology, Université du Québec à Montréal, Montréal, Québec, Canada
   - name: Mattan S. Ben-Shachar
     orcid: 0000-0002-4287-4801
-    affiliation: 2
+    affiliation: Independent Researcher, Ramat Gan, Israel
   - name: Indrajeet Patil
     orcid: 0000-0003-1995-6531
-    affiliation: 3
+    affiliation: Center for Humans and Machines, Max Planck Institute for Human Development, Berlin, Germany
   - name: Daniel Lüdecke
     orcid: 0000-0002-8895-3206
-    affiliation: 4
+    affiliation: Institute of Medical Sociology, University Medical Center Hamburg-Eppendorf, Germany
   - name: Brenton M. Wiernik
     orcid: 0000-0001-9560-6336
-    affiliation: 5
+    affiliation: Independent Researcher, Tampa, FL, USA
   - name: Dominique Makowski
     orcid: 0000-0001-5375-9967
-    affiliation: 6
+    affiliation: School of Psychology, University of Sussex, Brighton, UK
 affiliations:
   - index: 1
     name: Department of Psychology, Université du Québec à Montréal, Montréal, Québec, Canada
@@ -43,10 +43,8 @@ type: article
 status: submit
 date: 7 June 2023
 bibliography: paper.bib
-simplesummary: |
-  The *{performance}* package from the *easystats* ecosystem makes it easy to 
-  diagnose outliers in R and according to current best practices thanks to the 
-  `check_outiers()` function.
+abstract: |
+  Beyond the challenge of keeping up-to-date with current best practices regarding the diagnosis and treatment of outliers, an additional difficulty arises concerning the mathematical implementation of the recommended methods. Here, we provide an overview of current recommendations and best practices and demonstrate how they can easily and conveniently be implemented in the R statistical computing software, using the *{performance}* package of the *easystats* ecosystem. We cover univariate, multivariate, and model-based statistical outlier detection methods, their recommended threshold, standard output, and plotting methods. We conclude by reviewing the different theoretical types of outliers, whether to exclude or winsorize them, and the importance of transparency.
 keywords: |
   univariate outliers; multivariate outliers; robust detection methods; R; easystats
 acknowledgement: |
@@ -83,9 +81,7 @@ abbreviations:
     long: invariant coordinate selection
   - short: OSF
     long: Open Science Framework
-output: 
-  rticles::joss_article:
-    journal: "JOSE"
+output: rticles::arxiv_article
 csl: apa.csl
 ---
 
@@ -103,11 +99,7 @@ library(see)
 library(datawizard)
 ```
 
-# Summary
-
-Beyond the challenge of keeping up-to-date with current best practices regarding the diagnosis and treatment of outliers, an additional difficulty arises concerning the mathematical implementation of the recommended methods. Here, we provide an overview of current recommendations and best practices and demonstrate how they can easily and conveniently be implemented in the R statistical computing software, using the *{performance}* package of the *easystats* ecosystem. We cover univariate, multivariate, and model-based statistical outlier detection methods, their recommended threshold, standard output, and plotting methods. We conclude by reviewing the different theoretical types of outliers, whether to exclude or winsorize them, and the importance of transparency.
-
-# Statement of Need
+# Introduction
 
[…]
