+++ papers/preprint/paper.tex
@@ -1,0 +1,1008 @@
+\documentclass{article}
+
+\usepackage{arxiv}
+
+\usepackage[utf8]{inputenc} % allow utf-8 input
+\usepackage[T1]{fontenc}    % use 8-bit T1 fonts
+\usepackage{lmodern}        % https://github.com/rstudio/rticles/issues/343
+\usepackage{hyperref}       % hyperlinks
+\usepackage{url}            % simple URL typesetting
+\usepackage{booktabs}       % professional-quality tables
+\usepackage{amsfonts}       % blackboard math symbols
+\usepackage{nicefrac}       % compact symbols for 1/2, etc.
+\usepackage{microtype}      % microtypography
+\usepackage{graphicx}
+
+\title{Check your outliers! An introduction to identifying statistical
+outliers in R with \emph{easystats}}
+
+\author{
+    Rémi Thériault
+   \\
+    Department of Psychology, Université du Québec à Montréal, Montréal,
+Québec, Canada \\
+   \\
+  \texttt{} \\
+   \And
+    Mattan S. Ben-Shachar
+   \\
+    Independent Researcher, Ramat Gan, Israel \\
+   \\
+  \texttt{} \\
+   \And
+    Indrajeet Patil
+   \\
+    Center for Humans and Machines, Max Planck Institute for Human
+Development, Berlin, Germany \\
+   \\
+  \texttt{} \\
+   \And
+    Daniel Lüdecke
+   \\
+    Institute of Medical Sociology, University Medical Center
+Hamburg-Eppendorf, Germany \\
+   \\
+  \texttt{} \\
+   \And
+    Brenton M. Wiernik
+   \\
+    Independent Researcher, Tampa, FL, USA \\
+   \\
+  \texttt{} \\
+   \And
+    Dominique Makowski
+   \\
+    School of Psychology, University of Sussex, Brighton, UK \\
+   \\
+  \texttt{} \\
+  }
+
+% Pandoc syntax highlighting
+\usepackage{color}
+\usepackage{fancyvrb}
+\newcommand{\VerbBar}{|}
+\newcommand{\VERB}{\Verb[commandchars=\\\{\}]}
+\DefineVerbatimEnvironment{Highlighting}{Verbatim}{commandchars=\\\{\}}
+% Add ',fontsize=\small' for more characters per line
+\usepackage{framed}
+\definecolor{shadecolor}{RGB}{248,248,248}
+\newenvironment{Shaded}{\begin{snugshade}}{\end{snugshade}}
+\newcommand{\AlertTok}[1]{\textcolor[rgb]{0.94,0.16,0.16}{#1}}
+\newcommand{\AnnotationTok}[1]{\textcolor[rgb]{0.56,0.35,0.01}{\textbf{\textit{#1}}}}
+\newcommand{\AttributeTok}[1]{\textcolor[rgb]{0.13,0.29,0.53}{#1}}
+\newcommand{\BaseNTok}[1]{\textcolor[rgb]{0.00,0.00,0.81}{#1}}
+\newcommand{\BuiltInTok}[1]{#1}
+\newcommand{\CharTok}[1]{\textcolor[rgb]{0.31,0.60,0.02}{#1}}
+\newcommand{\CommentTok}[1]{\textcolor[rgb]{0.56,0.35,0.01}{\textit{#1}}}
[…]
