+++ pkgdown.yml
@@ -1,12 +1,12 @@
 pandoc: 3.1.8
 pkgdown: 2.0.7.9000
-pkgdown_sha: 170015984ae234da3028b2b083df757227fdd241
+pkgdown_sha: d5432a4e04655c6f9903196890f6bf166fbbd86b
 articles:
   check_model: check_model.html
   check_outliers: check_outliers.html
   compare: compare.html
   r2: r2.html
-last_built: 2023-10-17T09:37Z
+last_built: 2023-10-24T10:39Z
 urls:
   reference: https://easystats.github.io/performance/reference
   article: https://easystats.github.io/performance/articles
