### 2023-10-24 aff7628
CRAN submission 0.10.7
