### 2023-10-24 c5c7277
binom CI hack for binned_performance
