### 2023-10-24 2fb679a
still hacking on binned_residuals
