### 2023-10-25 bdc6ecb
lintr
