### 2023-10-26 eb4fafa
Improve `binned_residuals()` (#641)
