+++ tests/testthat/test-binned_residuals.R
@@ -1,10 +1,10 @@
 test_that("binned_residuals", {
   data(mtcars)
   model <- glm(vs ~ wt + mpg, data = mtcars, family = "binomial")
-  result <- binned_residuals(model)
+  result <- binned_residuals(model, ci_type = "gaussian", residuals = "response")
   expect_named(
     result,
-    c("xbar", "ybar", "n", "x.lo", "x.hi", "se", "ci_range", "CI_low", "CI_high", "group")
+    c("xbar", "ybar", "n", "x.lo", "x.hi", "se", "CI_low", "CI_high", "group")
   )
   expect_equal(
     result$xbar,
@@ -16,16 +16,21 @@ test_that("binned_residuals", {
     c(-0.03786, -0.09514, 0.07423, -0.07955, 0.28891, -0.13786),
     tolerance = 1e-4
   )
+  expect_equal(
+    result$CI_low,
+    c(-0.05686, -0.12331, -0.35077, -0.57683, 0.17916, -0.44147),
+    tolerance = 1e-4
+  )
 })
 
 
 test_that("binned_residuals, n_bins", {
   data(mtcars)
   model <- glm(vs ~ wt + mpg, data = mtcars, family = "binomial")
-  result <- binned_residuals(model, n_bins = 10)
+  result <- binned_residuals(model, ci_type = "gaussian", residuals = "response", n_bins = 10)
   expect_named(
     result,
-    c("xbar", "ybar", "n", "x.lo", "x.hi", "se", "ci_range", "CI_low", "CI_high", "group")
+    c("xbar", "ybar", "n", "x.lo", "x.hi", "se", "CI_low", "CI_high", "group")
   )
   expect_equal(
     result$xbar,
@@ -49,10 +54,10 @@ test_that("binned_residuals, n_bins", {
 test_that("binned_residuals, terms", {
   data(mtcars)
   model <- glm(vs ~ wt + mpg, data = mtcars, family = "binomial")
-  result <- binned_residuals(model, term = "mpg")
+  result <- binned_residuals(model, ci_type = "gaussian", residuals = "response", term = "mpg")
   expect_named(
     result,
-    c("xbar", "ybar", "n", "x.lo", "x.hi", "se", "ci_range", "CI_low", "CI_high", "group")
+    c("xbar", "ybar", "n", "x.lo", "x.hi", "se", "CI_low", "CI_high", "group")
   )
   expect_equal(
     result$xbar,
@@ -62,6 +67,91 @@ test_that("binned_residuals, terms", {
   expect_equal(
     result$ybar,
     c(-0.05435, -0.07866, 0.13925, -0.11861, 0.27763, -0.13786),
+    tolerance = 1e-4
+  )
+})
+
+
+test_that("binned_residuals, deviance residuals, gaussian CI", {
+  data(mtcars)
+  model <- glm(vs ~ wt + mpg, data = mtcars, family = "binomial")
+  result <- binned_residuals(model, residuals = "deviance", ci_type = "gaussian")
+  expect_named(
+    result,
+    c("xbar", "ybar", "n", "x.lo", "x.hi", "se", "CI_low", "CI_high", "group")
+  )
+  expect_equal(
+    result$xbar,
+    c(0.03786, 0.09514, 0.25911, 0.47955, 0.71109, 0.97119),
+    tolerance = 1e-4
+  )
+  expect_equal(
+    result$ybar,
+    c(-0.26905, -0.44334, 0.03763, -0.19917, 0.81563, -0.23399),
+    tolerance = 1e-4
+  )
+  expect_equal(
[…]
