### 2023-10-26 137a627
DHARMa implementation for new `check_residuals()` function Fixes #595
