+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Type: Package
 Package: performance
 Title: Assessment of Regression Models Performance
-Version: 0.10.6.3
+Version: 0.10.6.4
 Authors@R: 
     c(person(given = "Daniel",
              family = "Lüdecke",
@@ -91,6 +91,7 @@ Suggests:
     correlation,
     cplm,
     dbscan,
+    DHARMa,
     estimatr,
     fixest,
     flextable, 
