### 2023-10-26 4fc9274
draft function
