+++ man/simulate_residuals.Rd
@@ -1,0 +1,32 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/simulate_residuals.R
+\name{simulate_residuals}
+\alias{simulate_residuals}
+\title{Check model for (non-)constant error variance}
+\usage{
+simulate_residuals(x, ...)
+}
+\arguments{
+\item{x}{A model object.}
+
+\item{...}{Currently not used.}
+}
+\value{
+Simulated residuals.
+}
+\description{
+to do.
+}
+\details{
+Based on \code{\link[DHARMa:simulateResiduals]{DHARMa::simulateResiduals()}}.
+}
+\examples{
+m <<- lm(mpg ~ wt + cyl + gear + disp, data = mtcars)
+check_heteroscedasticity(m)
+
+# plot results
+if (require("see")) {
+  x <- check_heteroscedasticity(m)
+  plot(x)
+}
+}
