+++ NAMESPACE
@@ -576,6 +576,7 @@ export(r2_tjur)
 export(r2_xu)
 export(r2_zeroinflated)
 export(rmse)
+export(simulate_residuals)
 export(test_bf)
 export(test_likelihoodratio)
 export(test_lrt)
