### 2023-10-26 4c805bf
Don't prematurely return residuals in `simulate_residuals()`
