### 2023-10-26 bb1a2f3
draft `check_residuals()`
