### 2023-10-26 de272d7
Add (temporary?) vignette for basic simulated residuals workflow demonstration and discussion
