### 2023-10-26 a6a5afe
clarify why model fit is needed
