+++ man/check_residuals.Rd
@@ -1,0 +1,14 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/check_residuals.R
+\name{check_residuals}
+\alias{check_residuals}
+\title{Check uniformity of simulated residuals}
+\usage{
+check_residuals(x, ...)
+}
+\description{
+\code{check_residuals()} checks generalized linear (mixed) models for uniformity
+of randomized quantile residuals, which can be used to identify typical model
+misspecification problems, such as over/underdispersion, zero-inflation, and
+residual spatial and temporal autocorrelation.
+}
