+++ NAMESPACE
@@ -299,6 +299,7 @@ S3method(print,performance_pcp)
 S3method(print,performance_pp_check)
 S3method(print,performance_roc)
 S3method(print,performance_score)
+S3method(print,performance_simres)
 S3method(print,r2_bayes)
 S3method(print,r2_generic)
 S3method(print,r2_loo)
@@ -523,6 +524,7 @@ export(check_outliers)
 export(check_overdispersion)
 export(check_posterior_predictions)
 export(check_predictions)
+export(check_residuals)
 export(check_singularity)
 export(check_sphericity)
 export(check_sphericity_bartlett)
