### 2023-10-27 6ffa8b6
add method for check_normality
