+++ NAMESPACE
@@ -70,6 +70,7 @@ S3method(check_normality,htest)
 S3method(check_normality,lmerModLmerTest)
 S3method(check_normality,merMod)
 S3method(check_normality,numeric)
+S3method(check_normality,performance_simres)
 S3method(check_outliers,BFBayesFactor)
 S3method(check_outliers,character)
 S3method(check_outliers,data.frame)
