### 2023-10-27 991b483
add class attr for see::plot
