### 2023-10-27 50f1f3d
Add basic print output for `simulate_residuals()`
