### 2023-10-29 bfc8f8f
posterior predictive check for binomial glm with matrix response Fixes #644
