+++ tests/testthat/test-binned_residuals.R
@@ -155,3 +155,22 @@ test_that("binned_residuals, bootstrapped CI", {
     tolerance = 1e-4
   )
 })
+
+test_that("binned_residuals, msg for non-bernoulli", {
+  skip_on_cran()
+  tot <- rep(10, 100)
+  suc <- rbinom(100, prob = 0.9, size = tot)
+
+  dat <- data.frame(tot, suc)
+  dat$prop <- suc / tot
+  dat$x1 <- as.factor(sample(1:5, 100, replace = TRUE))
+
+  mod <- glm(prop ~ x1,
+    family = binomial,
+    data = df,
+    weights = tot
+  )
+
+  expect_message(binned_residuals(mod1), regex = "Using `ci_type = \"gaussian\"`")
+  expect_silent(binned_residuals(mod1, verbose = FALSE))
+})
