### 2023-10-29 50a0443
posterior predictive check for binomial glm with matrix response (#645)
