+++ pkgdown.yml
@@ -1,4 +1,4 @@
-pandoc: 3.1.8
+pandoc: 3.1.9
 pkgdown: 2.0.7.9000
 pkgdown_sha: d5432a4e04655c6f9903196890f6bf166fbbd86b
 articles:
@@ -6,7 +6,7 @@ articles:
   check_outliers: check_outliers.html
   compare: compare.html
   r2: r2.html
-last_built: 2023-10-28T07:18Z
+last_built: 2023-10-29T13:50Z
 urls:
   reference: https://easystats.github.io/performance/reference
   article: https://easystats.github.io/performance/articles
