+++ R/simulate_residuals.R
@@ -4,6 +4,7 @@
 #' @description to do.
 #'
 #' @param x A model object.
+#' @param iterations Number of simulations to run.
 #' @param ... Arguments passed on to [`DHARMa::simulateResiduals()`].
 #'
 #' @return Simulated residuals.
@@ -19,12 +20,12 @@
 #' - Dunn, P. K., & Smyth, G. K. (1996). Randomized Quantile Residuals. Journal
 #'   of Computational and Graphical Statistics, 5(3), 236. \doi{10.2307/1390802}
 #'
-#' @examples
+#' @examplesIf require("DHARMa")
 #' m <- lm(mpg ~ wt + cyl + gear + disp, data = mtcars)
 #' simulate_residuals(m)
 #'
 #' @export
-simulate_residuals <- function(x, ...) {
+simulate_residuals <- function(x, iterations = 250, ...) {
   insight::check_if_installed("DHARMa")
   # TODO (low priority): Note that DHARMa::simulateResiduals(x, ...) does its own checks for whether
   # or not the model passed to it is supported, do we want to use this or do our
@@ -34,7 +35,7 @@ simulate_residuals <- function(x, ...) {
   # extracting the residuals from it because the object contains important stuff
   # in it that we'll want to pass onto other functions later, such as passing
   # the fitted model into check_model().
-  out <- DHARMa::simulateResiduals(x, ...)
+  out <- DHARMa::simulateResiduals(x, n = iterations, plot = FALSE, ...)
   class(out) <- c("performance_simres", class(out))
   out
 }
@@ -46,8 +47,12 @@ print.performance_simres <- function(x, ...) {
   # TODO (low priority): We can probably just base this off of the print method
   # DHARMa uses, but with an easystats style. For now we can just stick with
   # DHARMa's method.
-  cat(
-    "Simulated residuals from a model of class", class(x$fittedModel),
-    "based on", x$nSim, "simulations."
+  msg <- paste0(
+    "Simulated residuals from a model of class `", class(x$fittedModel),
+    "` based on ", x$nSim, " simulations. Use `check_residuals()` to check ",
+    "uniformity of residuals. It is recommended to refer to `?DHARMa::simulateReisudals`",
+    " and `vignette(\"DHARMa\")` for more information about different settings",
+    " in particular situations or for particular models."
   )
+  cat(insight::format_message(msg))
 }
