+++ NAMESPACE
@@ -70,7 +70,6 @@ S3method(check_normality,htest)
 S3method(check_normality,lmerModLmerTest)
 S3method(check_normality,merMod)
 S3method(check_normality,numeric)
-S3method(check_normality,performance_simres)
 S3method(check_outliers,BFBayesFactor)
 S3method(check_outliers,character)
 S3method(check_outliers,data.frame)
@@ -106,6 +105,9 @@ S3method(check_overdispersion,poissonmfx)
 S3method(check_predictions,BFBayesFactor)
 S3method(check_predictions,default)
 S3method(check_predictions,lme)
+S3method(check_residuals,DHARMa)
+S3method(check_residuals,default)
+S3method(check_residuals,performance_simres)
 S3method(check_singularity,MixMod)
 S3method(check_singularity,clmm)
 S3method(check_singularity,cpglmm)
