### 2023-10-30 bd737f1
add plot method
