+++ NAMESPACE
@@ -265,6 +265,7 @@ S3method(plot,check_sphericity)
 S3method(plot,compare_performance)
 S3method(plot,performance_pp_check)
 S3method(plot,performance_roc)
+S3method(plot,performance_simres)
 S3method(plot,test_likelihoodratio)
 S3method(plot,test_performance)
 S3method(print,binned_residuals)
