+++ papers/preprint/cover_letter.Rmd
@@ -1,20 +1,25 @@
 ---
 output: pdf_document
+title: "Cover letter"
 ---
 
-Dear Dr. Vazire,
+Dear Dr. Steinley,
 
-We are pleased to submit this paper to *Collabra: Psychology*.
+We are pleased to submit this paper to *Psychological Methods* for the General Call for Papers: Tutorials in Psychological Methods.
 
 The paper, titled "Check your outliers! An introduction to identifying statistical outliers in R with *easystats*", provides an overview of current recommendations and best practices regarding the diagnosis and treatment of outliers, a common issue faced by researchers---and a potential source of scientific malpractice.
 
-It explains the key approaches, highlights recommendations, and shows how users can adopt them in their R analysis with a single function. The manuscript covers univariate, multivariate, and model-based statistical outlier detection methods, their recommended threshold, standard output, and plotting method, among other things.
+The methods / tutorial paper covers best practice recommendations on methods for outlier detection and treatment in statistical analyses, as well as their implementation (with code) in the easystats performance package in R. It explains the key approaches, highlights recommendations, and shows how users can adopt them in their R analysis with a single function. The manuscript covers univariate, multivariate, and model-based statistical outlier detection methods, their recommended threshold, standard output, and plotting method, among other things.
 
-Beyond acting like a concise review of outlier treatment procedures and practical tutorial, we also introduce a new outlier-detection method that relies on a consensus-based approach. In this sense, the paper fits well with the "Methodology and Research Practice in Psychology" section of the journal, as it essentially communicates to psychologists how to easily follow some of the best practices in the detection of statistical outlier using currently available open source and free software. This makes the manuscript relevant to data science, behavioural science, and good research and statistical practices more generally.
+Beyond acting like a concise review of outlier treatment procedures and practical tutorial, we also introduce a new outlier-detection method that relies on a consensus-based approach. In this sense, the paper fits well with the General Call for Papers: Tutorials in Psychological Methods, as the tutorial essentially communicates to psychologists how to easily follow some of the best practices in the detection of statistical outlier using currently available open source and free software. This makes the manuscript relevant to analyzing, understanding, and interpreting quantitative psychological data, psychological and behavioural science, and good research and statistical practices more generally.
 
-As Associated Editor, we would like to suggest Jeffrey Girard, as he is familiar with the *easystats* and R ecosystems, as well as good statistical practices. Additionally, we would like to request a streamlined review, as the manuscript has been rejected within the previous 365 days from the journal *Mathematics*. Accordingly, we provide in this submission a detailed letter that includes prior reviews, the decision letter, as well as how we addressed the reviewers' comments. We have in this regard integrated most of the changes suggested by the reviewers. Note that the previous editors and reviewers have not given their permission for their comments to be openly available at *Collabra: Psychology*. However, although an open review was requested, the reviewers did not sign their reviews.
+Additionally, we would like to suggest considering a streamlined review for this paper, as we submitted the paper to the *Journal of Open Source Education* (JOSE), a journal that publishes open source education materials and uses an open peer review model (hosted on GitHub). The paper went through several revisions with reviewers and was eventually enthusiastically recommended for acceptance by the reviewers and editor. The Editor in Chief then joined the conversation and declared the paper to be out of scope for the journal. The JOSE Editor in Chief’s comment is available here: https://github.com/openjournals/jose-reviews/issues/221#issuecomment-1767314739 and the open review history is discussed and linked above in the thread (including in the form of individual issues linked to the main package repository).
 
[…]
