+++ DESCRIPTION
@@ -70,7 +70,7 @@ Depends:
     R (>= 3.6)
 Imports:
     bayestestR (>= 0.13.1),
-    insight (>= 0.19.5),
+    insight (>= 0.19.6),
     datawizard (>= 0.9.0),
     methods,
     stats,
