### 2023-11-01 0f5c59f
add minority info
