+++ papers/preprint/cover_letter.Rmd
@@ -13,11 +13,9 @@ The methods / tutorial paper covers best practice recommendations on methods for
 
 Beyond acting like a concise review of outlier treatment procedures and practical tutorial, we also introduce a new outlier-detection method that relies on a consensus-based approach. In this sense, the paper fits well with the General Call for Papers: Tutorials in Psychological Methods, as the tutorial essentially communicates to psychologists how to easily follow some of the best practices in the detection of statistical outlier using currently available open source and free software. This makes the manuscript relevant to analyzing, understanding, and interpreting quantitative psychological data, psychological and behavioural science, and good research and statistical practices more generally.
 
-Additionally, we would like to suggest considering a streamlined review for this paper, as we submitted the paper to the *Journal of Open Source Education* (JOSE), a journal that publishes open source education materials and uses an open peer review model (hosted on GitHub). The paper went through several revisions with reviewers and was eventually enthusiastically recommended for acceptance by the reviewers and editor. The Editor in Chief then joined the conversation and declared the paper to be out of scope for the journal. The JOSE Editor in Chief’s comment is available here: https://github.com/openjournals/jose-reviews/issues/221#issuecomment-1767314739 and the open review history is discussed and linked above in the thread (including in the form of individual issues linked to the main package repository).
+Additionally, we would like to note that we submitted the paper to the *Journal of Open Source Education* (JOSE), a journal that publishes open source education materials and uses an open peer review model (hosted on GitHub). The paper went through several revisions with reviewers and was eventually enthusiastically recommended for acceptance by the reviewers and editor. However, the Editor in Chief then joined the conversation and declared the paper to be out of scope for the journal.
 
-Finally, regarding the General Call for Papers: Tutorials in Psychological Methods, we would like to highlight that: (a) our statistical outlier detection methods are of wide interest to a broad range of researchers in psychology; (b) our method paper communicates explicit guidelines, and includes reproducible examples using open computer code and data sets; (c) our proposed methodological procedures can be applied to several areas; (d) None of the current authors have previously published in *Psychological Methods*; (e) authors are from underrepresented groups and reflect diversity: The first and senior authors are early career researchers; only one of the six authors has English as their native language; only one of the six authors is from the USA; two authors belong to ethnic minorities.
-
-Authors from underrepresented groups (e.g., early career, women, ethnic minorities)
+Finally, regarding the General Call for Papers: Tutorials in Psychological Methods, we would like to highlight that: (a) our statistical outlier detection methods are of wide interest to a broad range of researchers in psychology; (b) our method paper communicates explicit guidelines, and includes reproducible examples using open computer code and data sets; (c) our proposed methodological procedures can be applied to several areas; (d) None of the current authors have previously published in *Psychological Methods*; (e) authors are from underrepresented groups and reflect diversity: The first and senior authors are early career researchers; only one of the six authors has English as their native language; only one of the six authors is from the USA; two authors belong to ethnic minorities; three authors are non-academics; and one author belongs to a sexual orientation minority.
 
 All authors have contributed substantially to the software and manuscript. All authors gave final approval to the manuscript and accept to be accountable. We have no conflicts of interest to disclose. We have also read the Transparency and Openness policy of the Editorial Policies of *Psychological Methods*.
 
