### 2023-11-01 a946468
drop JOSE paragraph
