+++ papers/preprint/cover_letter.Rmd
@@ -14,7 +14,7 @@ The methods / tutorial paper covers best practice recommendations on methods for
 
 Beyond acting like a concise review of outlier treatment procedures and practical tutorial, we also introduce a new outlier-detection method that relies on a consensus-based approach. In this sense, the paper fits well with the General Call for Papers: Tutorials in Psychological Methods, as the tutorial essentially communicates to psychologists how to easily follow some of the best practices in the detection of statistical outlier using currently available open source and free software. This makes the manuscript relevant to analyzing, understanding, and interpreting quantitative psychological data, psychological and behavioural science, and good research and statistical practices more generally.
 
-Finally, regarding the General Call for Papers: Tutorials in Psychological Methods, we would like to highlight that: (a) our statistical outlier detection methods are of wide interest to a broad range of researchers in psychology; (b) our method paper communicates explicit guidelines, and includes reproducible examples using open computer code and data sets; (c) our proposed methodological procedures can be applied to several areas; (d) None of the current authors have previously published in *Psychological Methods*; (e) authors are from underrepresented groups and reflect diversity: The first and senior authors are early career researchers; only one of the six authors has English as their native language; only one of the six authors is from the USA; two authors belong to ethnic minorities; three authors are non-academics; and one author belongs to a sexual orientation minority.
+Finally, regarding the General Call for Papers: Tutorials in Psychological Methods, we would like to highlight that: (a) Our statistical outlier detection methods are of wide interest to a broad range of researchers in psychology; (b) Our method paper communicates explicit guidelines, and includes reproducible examples using open computer code and data sets; (c) Our proposed methodological procedures can be applied to several areas; (d) None of the current authors have previously published in *Psychological Methods*; (e) The author group reflects diversity from several underrepresented groups: The first and senior authors are early career researchers; only one of the six authors has English as their native language; only one of the six authors is from the USA; two authors belong to ethnic minorities; three authors are non-academics; and one author belongs to a sexual orientation minority.
 
 All authors have contributed substantially to the software and manuscript. All authors gave final approval to the manuscript and accept to be accountable. We have no conflicts of interest to disclose. We have also read the Transparency and Openness policy of the Editorial Policies of *Psychological Methods*.
 
