### 2023-11-02 #221 @stats-tgeorge
Hey all,

First of all I want to apologize and take some credit for being uninformed. This is my second review as an coordinating editor and thus I don't know all of the JOSE fine details yet. I also miss-understood some of the documentation. I do want to encourage the authors to add to the vignette and and restructure as the EiC mentioned so we can review the new versions and get this published with JOSE. Let me know of additional questions and when you are ready to have new items reviewed!
