### 2023-11-03 d515869
add print method for `check_residuals()`
