+++ NAMESPACE
@@ -285,6 +285,7 @@ S3method(print,check_outliers)
 S3method(print,check_outliers_metafor)
 S3method(print,check_outliers_metagen)
 S3method(print,check_overdisp)
+S3method(print,check_residuals)
 S3method(print,check_sphericity)
 S3method(print,check_symmetry)
 S3method(print,check_zi)
