### 2023-11-07 e5fbd83
format paper for springer, add bit about reaction times
