+++ papers/preprint/paper.tex
@@ -1,61 +1,55 @@
-\documentclass{article}
+%Version 2.1 April 2023
+% See section 11 of the User Manual for version history
+%
+%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
+%%                                                                 %%
+%% Please do not use \input{...} to include other tex files.       %%
+%% Submit your LaTeX manuscript as one .tex document.              %%
+%%                                                                 %%
+%% All additional figures and files should be attached             %%
+%% separately and not embedded in the \TeX\ document itself.       %%
+%%                                                                 %%
+%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
 
-\usepackage{arxiv}
+\documentclass[sn-basic, lineno,pdflatex]{sn-jnl}
 
-\usepackage[utf8]{inputenc} % allow utf-8 input
-\usepackage[T1]{fontenc}    % use 8-bit T1 fonts
-\usepackage{lmodern}        % https://github.com/rstudio/rticles/issues/343
-\usepackage{hyperref}       % hyperlinks
-\usepackage{url}            % simple URL typesetting
-\usepackage{booktabs}       % professional-quality tables
-\usepackage{amsfonts}       % blackboard math symbols
-\usepackage{nicefrac}       % compact symbols for 1/2, etc.
-\usepackage{microtype}      % microtypography
-\usepackage{graphicx}
+%%%% Standard Packages
+%%<additional latex packages if required can be included here>
 
-\title{Check your outliers! An introduction to identifying statistical
-outliers in R with \emph{easystats}}
+\usepackage{graphicx}%
+\usepackage{multirow}%
+\usepackage{amsmath,amssymb,amsfonts}%
+\usepackage{amsthm}%
+\usepackage{mathrsfs}%
+\usepackage[title]{appendix}%
+\usepackage{xcolor}%
+\usepackage{textcomp}%
+\usepackage{manyfoot}%
+\usepackage{booktabs}%
+\usepackage{algorithm}%
+\usepackage{algorithmicx}%
+\usepackage{algpseudocode}%
+\usepackage{listings}%
+%%%%
 
-\author{
-    Rémi Thériault
-   \\
-    Department of Psychology, Université du Québec à Montréal, Montréal,
-Québec, Canada \\
-   \\
-  \texttt{} \\
-   \And
-    Mattan S. Ben-Shachar
-   \\
-    Independent Researcher, Ramat Gan, Israel \\
-   \\
-  \texttt{} \\
-   \And
-    Indrajeet Patil
-   \\
-    Center for Humans and Machines, Max Planck Institute for Human
-Development, Berlin, Germany \\
-   \\
-  \texttt{} \\
-   \And
-    Daniel Lüdecke
-   \\
-    Institute of Medical Sociology, University Medical Center
-Hamburg-Eppendorf, Germany \\
-   \\
-  \texttt{} \\
-   \And
-    Brenton M. Wiernik
-   \\
-    Independent Researcher, Tampa, FL, USA \\
-   \\
-  \texttt{} \\
-   \And
-    Dominique Makowski
-   \\
-    School of Psychology, University of Sussex, Brighton, UK \\
-   \\
-  \texttt{} \\
-  }
+%%%%%=============================================================================%%%%
+%%%%  Remarks: This template is provided to aid authors with the preparation
+%%%%  of original research articles intended for submission to journals published
+%%%%  by Springer Nature. The guidance has been prepared in partnership with
+%%%%  production teams to conform to Springer Nature technical requirements.
[…]
