+++ papers/preprint/paper.bib
@@ -1,3 +1,12 @@
+@article{miller2023outlier,
+  title={Outlier exclusion procedures for reaction time analysis: The cures are generally worse than the disease.},
+  author={Miller, Jeff},
+  journal={Journal of Experimental Psychology: General},
+  year={2023},
+  publisher={American Psychological Association},
+  doi = {10.1037/xge0001450}
+}
+
 @article{leys2019outliers,
   title = {How to Classify, Detect, and Manage Univariate and Multivariate Outliers, With Emphasis on Pre-Registration},author = {Leys, Christophe and Delacre, Marie and Mora, Youri L. and Lakens, Daniël and Ley, Christophe},
   journal = {International Review of Social Psychology},
@@ -44,8 +53,9 @@
 @software{easystatspackage,
   title = {{easystats}: Streamline Model Interpretation, Visualization, and Reporting},
   author = {Daniel Lüdecke and Dominique Makowski and Mattan S. Ben-Shachar and Indrajeet Patil and Brenton M. Wiernik and Etienne Bacher and Rémi Thériault},
-  date = {2023-02-04T22:06:06Z},
-  origdate = {2019-01-28T10:39:29Z},
+  year = {2023},
+  date = {2023-02-04},
+  origdate = {2019-01-28},
   url = {https://easystats.github.io/easystats/}
 }
 
