+++ papers/preprint/paper.Rmd
@@ -1,5 +1,6 @@
 ---
 title: "Check your outliers! An introduction to identifying statistical outliers in R with *easystats*"
+titlerunning: "Check your outliers"
 tags:
   - R
   - univariate outliers
@@ -7,37 +8,71 @@ tags:
   - robust detection methods
   - easystats
 authors:
-  - name: Rémi Thériault
+  # - name: Rémi Thériault
+  - firstname: Rémi
+    lastname: Thériault
     orcid: 0000-0003-4315-6788
-    affiliation: Department of Psychology, Université du Québec à Montréal, Montréal, Québec, Canada
-  - name: Mattan S. Ben-Shachar
+    affiliation: 1
+    corresponding: TRUE
+    email: theriault.remi@courrier.uqam.ca
+    
+  # - name: Mattan S. Ben-Shachar
+  - firstname: Mattan 
+    particle: S.
+    lastname: Ben-Shachar
     orcid: 0000-0002-4287-4801
-    affiliation: Independent Researcher, Ramat Gan, Israel
-  - name: Indrajeet Patil
+    affiliation: 2
+    
+  # - name: Indrajeet Patil
+  - firstname: Indrajeet
+    lastname: Patil
     orcid: 0000-0003-1995-6531
-    affiliation: Center for Humans and Machines, Max Planck Institute for Human Development, Berlin, Germany
-  - name: Daniel Lüdecke
+    affiliation: 3
+    
+  # - name: Daniel Lüdecke
+  - firstname: Daniel
+    lastname: Lüdecke
     orcid: 0000-0002-8895-3206
-    affiliation: Institute of Medical Sociology, University Medical Center Hamburg-Eppendorf, Germany
-  - name: Brenton M. Wiernik
+    affiliation: 4
+    
+  # - name: Brenton M. Wiernik
+  - firstname: Brenton
+    particle: M.
+    lastname: Wiernik
     orcid: 0000-0001-9560-6336
-    affiliation: Independent Researcher, Tampa, FL, USA
-  - name: Dominique Makowski
+    affiliation: 5
+    
+  # - name: Dominique Makowski
+  - firstname: Dominique
+    lastname: Makowski
     orcid: 0000-0001-5375-9967
-    affiliation: School of Psychology, University of Sussex, Brighton, UK
+    affiliation: 6
+
 affiliations:
-  - index: 1
-    name: Department of Psychology, Université du Québec à Montréal, Montréal, Québec, Canada
-  - index: 2
-    name: Independent Researcher, Ramat Gan, Israel
-  - index: 3
-    name: Center for Humans and Machines, Max Planck Institute for Human Development, Berlin, Germany
-  - index: 4
-    name: Institute of Medical Sociology, University Medical Center Hamburg-Eppendorf, Germany
-  - index: 5
-    name: Independent Researcher, Tampa, FL, USA
-  - index: 6
-    name: School of Psychology, University of Sussex, Brighton, UK
+  - number: 1
+    # name: Department of Psychology, Université du Québec à Montréal, Montréal, Québec, Canada
+    info:
+      orgname: Department of Psychology, Université du Québec à Montréal, Montréal, Québec, Canada
+  - number: 2
+    # name: Independent Researcher, Ramat Gan, Israel
+    info:
+      orgname: Independent Researcher, Ramat Gan, Israel
+  - number: 3
+    # name: Center for Humans and Machines, Max Planck Institute for Human Development, Berlin, Germany
+    info:
+      orgname: Center for Humans and Machines, Max Planck Institute for Human Development, Berlin, Germany
+  - number: 4
+    # name: Institute of Medical Sociology, University Medical Center Hamburg-Eppendorf, Germany
+    info:
[…]
