+++ papers/preprint/paper.bib
@@ -1,3 +1,58 @@
+@article{ratcliff2016diffusion,
+  title={Diffusion decision model: Current issues and history},
+  author={Ratcliff, Roger and Smith, Philip L and Brown, Scott D and McKoon, Gail},
+  journal={Trends in cognitive sciences},
+  volume={20},
+  number={4},
+  pages={260--281},
+  year={2016},
+  publisher={Elsevier},
+  doi = {10.1016/j.tics.2016.01.007}
+}
+
+@article{brown2008simplest,
+  title={The simplest complete model of choice response time: Linear ballistic accumulation},
+  author={Brown, Scott D and Heathcote, Andrew},
+  journal={Cognitive Psychology},
+  volume={57},
+  number={3},
+  pages={153--178},
+  year={2008},
+  publisher={Elsevier},
+  doi = {10.1016/j.cogpsych.2007.12.002}
+}
+
+@article{rouder2015lognormal,
+  title={The lognormal race: A cognitive-process model of choice and latency with desirable psychometric properties},
+  author={Rouder, Jeffrey N and Province, Jordan M and Morey, Richard D and Gomez, Pablo and Heathcote, Andrew},
+  journal={Psychometrika},
+  volume={80},
+  pages={491--513},
+  year={2015},
+  publisher={Springer},
+  doi = {10.1007/s11336-013-9396-3}
+}
+
+@article{anders2016shifted,
+  title={The shifted Wald distribution for response time data analysis.},
+  author={Anders, Royce and Alario, F and Van Maanen, Leendert and others},
+  journal={Psychological Methods},
+  volume={21},
+  number={3},
+  pages={309},
+  year={2016},
+  publisher={American Psychological Association},
+  doi = {10.1037/met0000066}
+}
+
+@article{schramm2019reaction,
+  title={Are reaction time transformations really beneficial?},
+  author={Schramm, Pele and Rouder, Jeffrey N},
+  year={2019},
+  journal={PsyArXiv},
+  doi = {10.31234/osf.io/9ksa6}
+}
+
 @article{miller2023outlier,
   title={Outlier exclusion procedures for reaction time analysis: The cures are generally worse than the disease.},
   author={Miller, Jeff},
