+++ papers/preprint/paper.tex
@@ -936,31 +936,55 @@ function of \emph{easystats} will help researchers engage in good
 research practices while providing a smooth outlier detection
 experience.
 
+\hypertarget{declarations}{%
+\section{Declarations}\label{declarations}}
+
+\hypertarget{funding-information}{%
+\subsection{Funding information}\label{funding-information}}
+
+This research received no external funding.
+
+\hypertarget{competing-interests}{%
+\subsection{Competing Interests}\label{competing-interests}}
+
+The authors declare no conflict of interest
+
+\hypertarget{availability-of-data-and-materials-data-transparency}{%
+\subsection{Availability of data and materials (data
+transparency)}\label{availability-of-data-and-materials-data-transparency}}
+
+This paper first appeared as a preprint
+(\url{https://doi.org/10.31234/osf.io/bu6nt}) and is also available as
+an online vignette at:
+\url{https://easystats.github.io/performance/articles/check_outliers}.
+All data used in this paper uses data included with base R.
+
+\hypertarget{code-availability}{%
+\subsection{Code availability}\label{code-availability}}
+
+The performance package is available at the package official website
+(\url{https://easystats.github.io/performance}), on CRAN
+(\url{https://cran.r-project.org/package=performance}), and on the
+R-Universe (\url{https://easystats.r-universe.dev/performance}). The
+source code is available on GitHub
+(\url{https://github.com/easystats/performance/}), and the package can
+be installed from CRAN with \texttt{install.packages("performance")}.
+
 \hypertarget{contributions}{%
-\subsubsection{Contributions}\label{contributions}}
+\subsection{Contributions}\label{contributions}}
 
 RT: Writing- Original draft preparation, Writing- Reviewing and Editing,
 Software. MSB-S, IP, DL, BMW, and DM: Writing- Reviewing and Editing,
 Software.
 
 \hypertarget{acknowledgements}{%
-\subsubsection{Acknowledgements}\label{acknowledgements}}
+\subsection{Acknowledgements}\label{acknowledgements}}
 
 \emph{\{performance\}} is part of the collaborative
 \href{https://github.com/easystats/easystats}{\emph{easystats}}
 ecosystem \citep{easystatspackage}. Thus, we thank all
 \href{https://github.com/orgs/easystats/people}{members of easystats},
 contributors, and users alike.
-
-\hypertarget{funding-information}{%
-\subsubsection{Funding information}\label{funding-information}}
-
-This research received no external funding.
-
-\hypertarget{competing-interests}{%
-\subsubsection{Competing Interests}\label{competing-interests}}
-
-The authors declare no conflict of interest
 
 \renewcommand\refname{References}
 \bibliography{paper.bib}
