+++ papers/preprint/cover_letter.Rmd
@@ -4,19 +4,19 @@ title: "Cover letter"
 date: "`r Sys.Date()`"
 ---
 
-Dear Dr. Steinley,
+Dear Dr. Brysbaert,
 
-We are pleased to submit this paper to *Psychological Methods* for the General Call for Papers: Tutorials in Psychological Methods.
+We are pleased to submit this paper to *Behavior Research Methods*.
 
 The paper, titled "Check your outliers! An introduction to identifying statistical outliers in R with *easystats*", provides an overview of current recommendations and best practices regarding the diagnosis and treatment of outliers, a common issue faced by researchers---and a potential source of scientific malpractice.
 
 The methods / tutorial paper covers best practice recommendations on methods for outlier detection and treatment in statistical analyses, as well as their implementation (with code) in the easystats performance package in R. It explains the key approaches, highlights recommendations, and shows how users can adopt them in their R analysis with a single function. The manuscript covers univariate, multivariate, and model-based statistical outlier detection methods, their recommended threshold, standard output, and plotting method, among other things.
 
-Beyond acting like a concise review of outlier treatment procedures and practical tutorial, we also introduce a new outlier-detection method that relies on a consensus-based approach. In this sense, the paper fits well with the General Call for Papers: Tutorials in Psychological Methods, as the tutorial essentially communicates to psychologists how to easily follow some of the best practices in the detection of statistical outlier using currently available open source and free software. This makes the manuscript relevant to analyzing, understanding, and interpreting quantitative psychological data, psychological and behavioural science, and good research and statistical practices more generally.
+Beyond acting like a concise review of outlier treatment procedures and practical tutorial, we also introduce a new outlier-detection method that relies on a consensus-based approach. In this sense, the paper fits well with *Behavior Research Methods*, as the tutorial essentially communicates to psychologists how to easily follow some of the best practices in the detection of statistical outlier using currently available open source and free software. This makes the manuscript relevant to analyzing, understanding, and interpreting quantitative psychological data, psychological and behavioural science, and good research and statistical practices more generally.
 
-Finally, regarding the General Call for Papers: Tutorials in Psychological Methods, we would like to highlight that: (a) Our statistical outlier detection methods are of wide interest to a broad range of researchers in psychology; (b) Our method paper communicates explicit guidelines, and includes reproducible examples using open computer code and data sets; (c) Our proposed methodological procedures can be applied to several areas; (d) None of the current authors have previously published in *Psychological Methods*; (e) The author group reflects diversity from several underrepresented groups: The first and senior authors are early career researchers; only one of the six authors has English as their native language; only one of the six authors is from the USA; two authors belong to ethnic minorities; three authors are non-academics; and one author belongs to a sexual orientation minority.
+Finally, as per our personal communication, we are happy to share that we have added a section covering the management of outliers for reaction times, since much research in psychology relies on this type of task. We specifically discuss the Miller (2023) paper, and others, in the second and third paragraphs of Section 3 ("Handling Outliers"). We hope this is satisfying and are looking forward to further feedback.
 
-All authors have contributed substantially to the software and manuscript. All authors gave final approval to the manuscript and accept to be accountable. We have no conflicts of interest to disclose. We have also read the Transparency and Openness policy of the Editorial Policies of *Psychological Methods*.
+All authors have contributed substantially to the software and manuscript. All authors gave final approval to the manuscript and accept to be accountable. We have no conflicts of interest to disclose.
 
 Thank you for considering our submission.
 
