### 2023-11-21 ea28c89
performance::r2_nakagawa() and r.squaredGLMM() give different values for Gaussian glmmTMB models without random effects Fixes #652
