+++ R/check_clusterstructure.R
@@ -114,28 +114,25 @@ plot.check_clusterstructure <- function(x, ...) {
     p[, i] <- stats::runif(n, min = cc[i], max = d[i])
   }
   k <- round(stats::runif(n, 1, nrow(x)))
-  q <- as.matrix(x[k, ])
+  qq <- as.matrix(x[k, ])
   distp <- rep(0, nrow(x))
-  # distq=rep(0,nrow(x)-1)
   distq <- 0
   minp <- rep(0, n)
   minq <- rep(0, n)
   for (i in 1:n) {
     distp[1] <- stats::dist(rbind(p[i, ], x[1, ]), method = distance)
-    minqi <- stats::dist(rbind(q[i, ], x[1, ]), method = distance)
+    minqi <- stats::dist(rbind(qq[i, ], x[1, ]), method = distance)
     for (j in 2:nrow(x)) {
       distp[j] <- stats::dist(rbind(p[i, ], x[j, ]), method = distance)
-      error <- q[i, ] - x[j, ]
+      error <- qq[i, ] - x[j, ]
       if (sum(abs(error)) != 0) {
-        # distq[j]<-stats::dist(rbind(q[i,],x[j,]))
-        distq <- stats::dist(rbind(q[i, ], x[j, ]), method = distance)
+        distq <- stats::dist(rbind(qq[i, ], x[j, ]), method = distance)
         if (distq < minqi) {
           minqi <- distq
         }
       }
     }
     minp[i] <- min(distp)
-    # minq[i]<-apply(distq,1,min)
     minq[i] <- minqi
   }
   sum(minq) / (sum(minp) + sum(minq))
