### 2023-11-21 8329231
add more r2 methods for glmmTMB
