+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Type: Package
 Package: performance
 Title: Assessment of Regression Models Performance
-Version: 0.10.8.2
+Version: 0.10.8.3
 Authors@R: 
     c(person(given = "Daniel",
              family = "Lüdecke",
