### 2023-11-23 ded5c5d
add more r2 to non-mixed glmmTMB
