+++ tests/testthat/test-r2.R
@@ -66,7 +66,7 @@ withr::with_environment(
     out2 <- r2(m2)
     expect_equal(out[[1]], out2[[1]], tolerance = 1e-3, ignore_attr = TRUE)
     # poisson --------------------------------------------------------------
-    d <- data.frame(
+    d <<- data.frame(
       counts = c(18, 17, 15, 20, 10, 20, 25, 13, 12),
       outcome = gl(3, 1, 9),
       treatment = gl(3, 3)
@@ -97,7 +97,7 @@ withr::with_environment(
     out2 <- r2(m2)
     expect_equal(out[[1]], out2[[1]], tolerance = 1e-3, ignore_attr = TRUE)
     # Gamma --------------------------------------------------------------
-    clotting <- data.frame(
+    clotting <<- data.frame(
       u = c(5, 10, 15, 20, 30, 40, 60, 80, 100),
       lot1 = c(118, 58, 42, 35, 27, 25, 21, 19, 18),
       lot2 = c(69, 35, 26, 21, 18, 16, 13, 12, 12)
