+++ tests/testthat/test-r2.R
@@ -40,23 +40,74 @@ test_that("r2 glm, ci", {
   )
 })
 
-test_that("r2 glmmTMB, no ranef", {
-  skip_if_not_installed("glmmTMB")
-  data(Owls, package = "glmmTMB")
-  m <- glmmTMB::glmmTMB(NegPerChick ~ BroodSize + ArrivalTime, data = Owls)
-  out <- r2(m)
-  expect_equal(out$R2, 0.05597288, tolerance = 1e-3, ignore_attr = TRUE)
-  # validate against lm
-  m2 <- lm(NegPerChick ~ BroodSize + ArrivalTime, data = Owls)
-  out2 <- r2(m2)
-  expect_equal(out$R2, out2$R2, tolerance = 1e-3, ignore_attr = TRUE)
-  # binomial
-  data(mtcars)
-  m <- glmmTMB::glmmTMB(am ~ mpg, data = mtcars, family = binomial())
-  out <- r2(m)
-  expect_equal(out[[1]], 0.3677326, tolerance = 1e-3, ignore_attr = TRUE)
-  # validate against glm
-  m2 <- glm(am ~ mpg, data = mtcars, family = binomial())
-  out2 <- r2(m2)
-  expect_equal(out[[1]], out[[1]], tolerance = 1e-3, ignore_attr = TRUE)
-})
+# glmmTMB, non-mixed --------------------------------------------------------
+
+skip_if_not_installed("withr")
+withr::with_environment(
+  new.env(),
+  test_that("r2 glmmTMB, no ranef", {
+    skip_if_not_installed("glmmTMB")
+    data(Owls, package = "glmmTMB")
+    # linear ---------------------------------------------------------------
+    m <- glmmTMB::glmmTMB(NegPerChick ~ BroodSize + ArrivalTime, data = Owls)
+    out <- r2(m)
+    expect_equal(out$R2, 0.05597288, tolerance = 1e-3, ignore_attr = TRUE)
+    # validate against lm
+    m2 <- lm(NegPerChick ~ BroodSize + ArrivalTime, data = Owls)
+    out2 <- r2(m2)
+    expect_equal(out$R2, out2$R2, tolerance = 1e-3, ignore_attr = TRUE)
+    # binomial -------------------------------------------------------------
+    data(mtcars)
+    m <- glmmTMB::glmmTMB(am ~ mpg, data = mtcars, family = binomial())
+    out <- r2(m)
+    expect_equal(out[[1]], 0.3677326, tolerance = 1e-3, ignore_attr = TRUE)
+    # validate against glm
+    m2 <- glm(am ~ mpg, data = mtcars, family = binomial())
+    out2 <- r2(m2)
+    expect_equal(out[[1]], out2[[1]], tolerance = 1e-3, ignore_attr = TRUE)
+    # poisson --------------------------------------------------------------
+    d <<- data.frame(
+      counts = c(18, 17, 15, 20, 10, 20, 25, 13, 12),
+      outcome = gl(3, 1, 9),
+      treatment = gl(3, 3)
+    )
+    m <- glmmTMB::glmmTMB(counts ~ outcome + treatment, family = poisson(), data = d)
+    out <- r2(m)
+    expect_equal(out[[1]], 0.6571698, tolerance = 1e-3, ignore_attr = TRUE)
+    # validate against glm
+    m2 <- glm(counts ~ outcome + treatment, family = poisson(), data = d)
+    out2 <- r2(m2)
+    expect_equal(out[[1]], out2[[1]], tolerance = 1e-3, ignore_attr = TRUE)
+    # zero-inflated --------------------------------------------------------------
[…]
