+++ R/r2_nagelkerke.R
@@ -102,7 +102,8 @@ r2_nagelkerke.glmmTMB <- function(model, verbose = TRUE, ...) {
     return(NULL)
   }
 
-  null_dev <- stats::deviance(insight::null_model(model))
+  null_mod <- suppressWarnings(insight::null_model(model))
+  null_dev <- stats::deviance(null_mod)
   r2cox <- (1 - exp((dev - null_dev) / insight::n_obs(model, disaggregate = TRUE)))
 
   if (is.na(r2cox) || is.null(r2cox)) {
