+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Type: Package
 Package: performance
 Title: Assessment of Regression Models Performance
-Version: 0.10.8.4
+Version: 0.10.8.5
 Authors@R: 
     c(person(given = "Daniel",
              family = "Lüdecke",
@@ -135,10 +135,10 @@ Suggests:
     rstanarm,
     rstantools,
     sandwich,
-    see (>= 0.7.5),
+    see (>= 0.8.1),
     survey,
     survival,
-    testthat,
+    testthat (>= 3.2.0),
     tweedie,
     VGAM,
     withr
