### 2023-11-29 654a1e2
Simple glms no longer supported? Fixes #659
