### 2023-11-29 d0f7bd6
Simple glms no longer supported? (#660)
