+++ tests/testthat/test-binned_residuals.R
@@ -164,7 +164,7 @@ test_that("binned_residuals, msg for non-bernoulli", {
 
   dat <- data.frame(tot, suc)
   dat$prop <- suc / tot
-  dat$x1 <- as.factor(sample(1:5, 100, replace = TRUE))
+  dat$x1 <- as.factor(sample.int(5, 100, replace = TRUE))
 
   mod <- glm(prop ~ x1,
     family = binomial,
@@ -174,4 +174,70 @@ test_that("binned_residuals, msg for non-bernoulli", {
 
   expect_message(binned_residuals(mod), regex = "Using `ci_type = \"gaussian\"`")
   expect_silent(binned_residuals(mod, verbose = FALSE))
+})
+
+test_that("binned_residuals, empty bins", {
+  eel <- data.frame(
+    cured_bin = c(
+      1, 1, 1, 0, 0, 0, 1, 0, 0, 1, 0,
+      0, 0, 1, 1, 1, 0, 1, 0, 0, 1, 0, 0, 1, 0, 0, 0, 0, 0, 0, 1, 0,
+      0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0,
+      0, 1, 0, 0, 1, 0, 0, 0, 1, 0, 1, 0, 1, 1, 0, 1, 1, 0, 1, 1, 0,
+      0, 0, 0, 1, 0, 1, 0, 1, 0, 0, 1, 0, 0, 1, 0, 1, 0, 1, 0, 0, 1,
+      0, 0, 1, 1, 0, 1, 1, 0, 0, 1, 0, 1, 1, 0, 0, 1, 1, 0
+    ),
+    intervention = c(
+      "No treatment",
+      "No treatment", "No treatment", "No treatment", "Intervention",
+      "No treatment", "Intervention", "Intervention", "No treatment",
+      "No treatment", "Intervention", "No treatment", "No treatment",
+      "Intervention", "No treatment", "No treatment", "Intervention",
+      "Intervention", "Intervention", "Intervention", "No treatment",
+      "Intervention", "Intervention", "No treatment", "Intervention",
+      "Intervention", "No treatment", "No treatment", "Intervention",
+      "Intervention", "No treatment", "No treatment", "Intervention",
+      "Intervention", "Intervention", "No treatment", "No treatment",
+      "Intervention", "No treatment", "Intervention", "No treatment",
+      "Intervention", "Intervention", "Intervention", "No treatment",
+      "No treatment", "No treatment", "Intervention", "Intervention",
+      "No treatment", "Intervention", "Intervention", "Intervention",
+      "No treatment", "No treatment", "Intervention", "Intervention",
+      "No treatment", "Intervention", "Intervention", "No treatment",
+      "No treatment", "No treatment", "Intervention", "Intervention",
+      "No treatment", "No treatment", "No treatment", "No treatment",
+      "No treatment", "Intervention", "No treatment", "Intervention",
+      "Intervention", "Intervention", "No treatment", "Intervention",
+      "Intervention", "No treatment", "Intervention", "No treatment",
+      "No treatment", "Intervention", "Intervention", "Intervention",
+      "Intervention", "No treatment", "Intervention", "Intervention",
[…]
