### 2023-12-13 #221 @rempsyc
Dear reviewers and editors,

First, I would like to thank you all for contributing your reflections and viewpoints on this matter. We have decided to withdraw this publication from JOSE to submit to a different journal.

Ultimately, we realize that the format we were envisioning (i.e., the paper itself being the resource) is not compatible with the JOSE model (i.e., the paper simply linking to the resource), as we want the paper to be cited for its best practice recommendations in itself like a regular paper (people would probably not cite a short paper describing a vignette on good practices). This is due to a misunderstanding on our part; we missed that the paper itself could not contain educational material. We recognize that our desired outcome would break the JOSE model, so we believe it is a better outcome for both us authors and JOSE to proceed with the paper withdrawal.

I would like to reassure the reviewers and editors that their work has not been in vain, since it has contributed to making the paper (and associated resources) stronger. We thank you for that. Because of our paper withdrawal, I think this issue can be closed. Thank you for your understanding, and I will let you know if this gets published elsewhere :)
