+++ papers/preprint/paper.bib
@@ -1,7 +1,7 @@
 @article{ratcliff2016diffusion,
   title={Diffusion decision model: Current issues and history},
   author={Ratcliff, Roger and Smith, Philip L and Brown, Scott D and McKoon, Gail},
-  journal={Trends in cognitive sciences},
+  journal={Trends in Cognitive Sciences},
   volume={20},
   number={4},
   pages={260--281},
@@ -34,7 +34,7 @@
 }
 
 @article{anders2016shifted,
-  title={The shifted Wald distribution for response time data analysis.},
+  title={The shifted {Wald} distribution for response time data analysis.},
   author={Anders, Royce and Alario, F and Van Maanen, Leendert and others},
   journal={Psychological Methods},
   volume={21},
@@ -82,7 +82,7 @@
 }
 
 @article{leys2018outliers,
-  title = {Detecting multivariate outliers: Use a robust variant of the Mahalanobis distance},
+  title = {Detecting multivariate outliers: Use a robust variant of the {Mahalanobis} distance},
   journal = {Journal of Experimental Social Psychology},
   volume = {74},
   pages = {150-156},
@@ -162,7 +162,9 @@
   doi = {10.2307/2528963},
   journal = {Biometrics},
   pages = {81-124},
-  year = {1972}
+  year = {1972},
+  volume = {28},
+  number = {1}
 }
 
 @article{hubert2018mcd,
@@ -201,7 +203,7 @@
 @article{ratcliff1993methods,
   title={Methods for dealing with reaction time outliers.},
   author={Ratcliff, Roger},
-  journal={Psychological bulletin},
+  journal={Psychological Bulletin},
   volume={114},
   number={3},
   pages={510},
@@ -215,6 +217,17 @@
   author={McElreath, Richard},
   year={2020},
   publisher={CRC press}
+}
+
+@article{theriault2023rempsyc,
+  title={{rempsyc}: Convenience functions for psychology},
+  author={Thériault, Rémi},
+  journal={Journal of Open Source Software},
+  volume={8},
+  number={87},
+  pages={5466},
+  year={2023},
+  doi = {10.1037/0033-2909.114.3.510}
 }
 
 @Manual{rcore,
