+++ papers/preprint/paper.Rmd
@@ -117,7 +117,8 @@ abbreviations:
   - short: OSF
     long: Open Science Framework
 # output: rticles::arxiv_article
-output: rticles::springer_article
+# output: rticles::springer_article
+output: word_document
 journal: "BRM"
 classoptions: sn-basic, lineno
 csl: apa.csl
@@ -139,7 +140,7 @@ library(datawizard)
 
 # Introduction
 
-Real-life data often contain observations that can be considered *abnormal* when compared to the main population. The cause of it can be hard to assess and the boundaries of "abnormal", difficult to define---they may belong to a different distribution (originating from a different generative process) or simply be extreme cases, statistically rare but not impossible.
+Real-life data often contain observations that can be considered *abnormal* when compared to the main population. The cause of it can be hard to assess and the boundaries of "abnormal" difficult to define---they may belong to a different distribution (originating from a different generative process) or simply be extreme cases, statistically rare but not impossible.
 
 Nonetheless, the improper handling of these outliers can substantially affect statistical model estimations, biasing effect estimations and weakening the models' predictive performance. It is thus essential to address this problem in a thoughtful manner. Yet, despite the existence of established recommendations and guidelines, many researchers still do not treat outliers in a consistent manner, or do so using inappropriate strategies [@simmons2011false; @leys2013outliers].
 
@@ -190,7 +191,7 @@ which(outliers)
 data_clean <- data[-which(outliers), ]
 ```
 
-All `check_outliers()` output objects possess a `plot()` method, meaning it is also possible to visualize the outliers using the generic `plot()` function on the resulting outlier object after loading the {see} package (Figure 1).
+All `check_outliers()` output objects possess a `plot()` method, meaning it is also possible to visualize all observations in a way that highlights the outliers using the generic `plot()` function on the resulting outlier object after loading the {see} package (Figure 1).
 
 ```{r univariate, eval=FALSE}
 library(see)
@@ -198,7 +199,11 @@ library(see)
 plot(outliers)
 ```
 
-```{r univariate_implicit, fig.cap = "Visual depiction of outliers using the robust z-score method. The distance represents an aggregate score for variables mpg, cyl, disp, and hp.", echo=FALSE}
+**Figure 1**
+
+_Visual depiction of outliers using the robust z-score method_
+
+```{r univariate_implicit, fig.cap = "Note. The distance represents an aggregate score for variables mpg, cyl, disp, and hp.", echo=FALSE}
 library(see)
 
 plot(outliers) + 
@@ -224,13 +229,17 @@ outliers <- check_outliers(data, method = "mcd")
 outliers
 ```
 
-Here, we detected 9 multivariate outliers (i.e,. when looking at all variables of our dataset together).
+Here, we detected 9 multivariate outliers (i.e,. when looking at all variables of our dataset together). We can see the result in Figure 2.
+
+**Figure 2**
+
+_Visual depiction of outliers using the Minimum Covariance Determinant (MCD) method_
 
 ```{r multivariate_plot, eval=FALSE}
 plot(outliers)
 ```
 
-```{r multivariate_implicit, fig.cap = "Visual depiction of outliers using the Minimum Covariance Determinant (MCD) method, a robust version of the Mahalanobis distance. The distance represents the MCD scores for variables mpg, cyl, disp, and hp.", echo=FALSE}
+```{r multivariate_implicit, fig.cap = "Note. The Minimum Covariance Determinant (MCD) method is a robust version of the Mahalanobis distance. The distance represents the MCD scores for variables mpg, cyl, disp, and hp.", echo=FALSE}
 plot(outliers) + 
   ggplot2::theme(axis.text.x = ggplot2::element_text(
[…]
