### 2024-03-29 #221 @rempsyc
Dear editors and reviewers, I have the wonderful news that this week the paper was accepted in _Behavior Research Methods_. This has been a long ride and we have faced several journal rejections along the way, but it has been worth it. On behalf of the team, I would like to thank everyone involved in this constructive process that contributed to making the paper and related materials stronger. Thank you 🙏

> Thériault, R., Ben-Shachar, M. S., Patil, I., Lüdecke, D., Wiernik, B. M., & Makowski, D. (2024). Check your outliers﻿! An introduction to identifying statistical outliers in R with easystats. _Behavior Research Methods_, 1-11. https://doi.org/10.3758/s13428-024-02356-w
